Fix compilation errors in OpenTK.AL.
