+++ src/OpenTK.OpenAL/Native/AL/AL.cs
@@ -1,6 +1,6 @@
 ﻿/* AlFunctions.cs
  * C header: \OpenAL 1.1 SDK\include\Al.h
- * Spec: http://www.openal.org/openal_webstf/specs/OpenAL11Specification.pdf
+ * Spec: http://www.openal.org/openal_webstf/specs/Open<see cref="AL11Specification.pdf"/>
  * Copyright (c) 2008 Christoph Brandtner and Stefanos Apostolopoulos
  * See license.txt for license details
  * http://www.OpenTK.net */
@@ -11,56 +11,6 @@ using System.Security;
 using OpenTK.Core.Utility;
 using OpenTK.Mathematics;
 
-/* Type Mapping
-// 8-bit boolean
-typedef char ALboolean;
- * bool
-// character
-typedef char ALchar;
- * byte
-// signed 8-bit 2's complement integer
-typedef char ALbyte;
- * byte
-
-// unsigned 8-bit integer
-typedef unsigned char ALubyte;
- * byte
-
-// signed 16-bit 2's complement integer
-typedef short ALshort;
- * short
-
-// unsigned 16-bit integer
-typedef unsigned short ALushort;
- * ushort
-
-// unsigned 32-bit integer
-typedef unsigned int ALuint;
- * uint
-
-// signed 32-bit 2's complement integer
-typedef int ALint;
- * int
-// non-negative 32-bit binary integer size
-typedef int ALsizei;
- * int
-// enumerated 32-bit value
-typedef int ALenum;
- * int
-
-// 32-bit IEEE754 floating-point
-typedef float ALfloat;
- * float
-
-// 64-bit IEEE754 floating-point
-typedef double ALdouble;
- * double
-
-// void type (for opaque pointers only)
-typedef void ALvoid;
- * void
-*/
-
 namespace OpenTK.OpenAL.Native
 {
     /// <summary>
@@ -68,7 +18,10 @@ namespace OpenTK.OpenAL.Native
     /// </summary>
     public static class AL
     {
-        internal const string Lib = "openal32.dll";
+        /// <summary>
+        /// The name of the library.
+        /// </summary>
+        internal const string Library = "openal32.dll";
         private const CallingConvention Style = CallingConvention.Cdecl;
 
         /// <summary>
@@ -76,40 +29,35 @@ namespace OpenTK.OpenAL.Native
         /// used with this function, but it may be used by an extension.
         /// </summary>
         /// <param name="capability">The name of a capability to enable.</param>
-        [DllImport(Lib, EntryPoint = "alEnable", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alEnable", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void Enable(ALCapability capability);
-        // AL_API void AL_APIENTRY alEnable( ALenum capability );
 
         /// <summary>
         /// This function disables a feature of the OpenAL driver.
         /// </summary>
         /// <param name="capability">The name of a capability to disable.</param>
-        [DllImport(Lib, EntryPoint = "alDisable", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alDisable", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void Disable(ALCapability capability);
-        // AL_API void AL_APIENTRY alDisable( ALenum capability );
 
         /// <summary>
         /// This function returns a boolean indicating if a specific feature is enabled in the OpenAL driver.
         /// </summary>
         /// <param name="capability">The name of a capability to enable.</param>
         /// <returns>True if enabled, False if disabled.</returns>
-        [DllImport(Lib, EntryPoint = "alIsEnabled", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alIsEnabled", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern bool IsEnabled(ALCapability capability);
-        // AL_API ALboolean AL_APIENTRY alIsEnabled( ALenum capability );
 
-        [DllImport(Lib, EntryPoint = "alGetString", ExactSpelling = true, CallingConvention = Style,
-            CharSet = CharSet.Ansi)]
+        [DllImport(Library, EntryPoint = "alGetString", ExactSpelling = true, CallingConvention = Style, CharSet = CharSet.Ansi)]
         [SuppressUnmanagedCodeSecurity]
-        private static extern IntPtr GetStringPrivate(ALGetString param); // accepts the enums AlError, AlContextString
-        // AL_API const ALchar* AL_APIENTRY alGetString( ALenum param );
+        private static extern IntPtr GetStringPrivate(ALGetString param);
 
         /// <summary>
         /// This function retrieves an OpenAL string property.
         /// </summary>
-        /// <param name="param">The property to be returned: Vendor, Version, Renderer and Extensions</param>
+        /// <param name="param">The property to be returned: Vendor, Version, Renderer and Extensions.</param>
         /// <returns>Returns a pointer to a null-terminated string.</returns>
         public static string Get(ALGetString param)
         {
@@ -119,61 +67,30 @@ namespace OpenTK.OpenAL.Native
         /// <summary>
         /// This function retrieves an OpenAL string property.
         /// </summary>
-        /// <param name="param">The human-readable errorstring to be returned.</param>
+        /// <param name="param">The human-readable error string to be returned.</param>
         /// <returns>Returns a pointer to a null-terminated string.</returns>
         public static string GetErrorString(ALError param)
         {
             return Marshal.PtrToStringAnsi(GetStringPrivate((ALGetString)param));
         }
 
-        /* no functions return more than 1 result ..
-        // AL_API void AL_APIENTRY alGetBooleanv( ALenum param, ALboolean* buffer );
-        // AL_API void AL_APIENTRY alGetIntegerv( ALenum param, ALint* buffer );
-        // AL_API void AL_APIENTRY alGetFloatv( ALenum param, ALfloat* buffer );
-        // AL_API void AL_APIENTRY alGetDoublev( ALenum param, ALdouble* buffer );
-        */
-
-        /* disabled due to no token using it
-        /// <summary>
-/// This function returns a boolean OpenAL state.
-/// </summary>
-        /// <param name="param">the state to be queried: AL_DOPPLER_FACTOR, AL_SPEED_OF_SOUND, AL_DISTANCE_MODEL</param>
-        /// <returns>The boolean state described by param will be returned.</returns>
-        [DllImport( AL.Lib, EntryPoint = "alGetBoolean", ExactSpelling = true, CallingConvention = AL.Style ), SuppressUnmanagedCodeSecurity( )]
-        public static extern bool Get( ALGetBoolean param );
-        // AL_API ALboolean AL_APIENTRY alGetBoolean( ALenum param );
-        */
-
         /// <summary>
         /// This function returns an integer OpenAL state.
         /// </summary>
         /// <param name="param">the state to be queried: DistanceModel.</param>
         /// <returns>The integer state described by param will be returned.</returns>
-        [DllImport(Lib, EntryPoint = "alGetInteger", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGetInteger", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern int Get(ALGetInteger param);
-        // AL_API ALint AL_APIENTRY alGetInteger( ALenum param );
 
         /// <summary>
         /// This function returns a floating-point OpenAL state.
         /// </summary>
         /// <param name="param">the state to be queried: DopplerFactor, SpeedOfSound.</param>
         /// <returns>The floating-point state described by param will be returned.</returns>
-        [DllImport(Lib, EntryPoint = "alGetFloat", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGetFloat", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern float Get(ALGetFloat param);
-        // AL_API ALfloat AL_APIENTRY alGetFloat( ALenum param );
-
-        /* disabled due to no token using it
-        /// <summary>
-/// This function returns a double-precision floating-point OpenAL state.
-/// </summary>
-        /// <param name="param">the state to be queried: AL_DOPPLER_FACTOR, AL_SPEED_OF_SOUND, AL_DISTANCE_MODEL</param>
-        /// <returns>The double value described by param will be returned.</returns>
-        [DllImport( AL.Lib, EntryPoint = "alGetDouble", ExactSpelling = true, CallingConvention = AL.Style ), SuppressUnmanagedCodeSecurity( )]
-        public static extern double Get( ALGetDouble param );
-        // AL_API ALdouble AL_APIENTRY alGetDouble( ALenum param );
-        */
 
         /// <summary>
         /// Error support. Obtain the most recent error generated in the AL state machine. When an error is detected by
@@ -182,27 +99,24 @@ namespace OpenTK.OpenAL.Native
         /// its code.
         /// </summary>
         /// <returns>
-        /// The first error that occured. can be used with AL.GetString. Returns an Alenum representing the error state.
+        /// The first error that occured. can be used with <see cref="AL.GetString"/>. Returns an Alenum representing the error state.
         /// When an OpenAL error occurs, the error state is set and will not be changed until the error state is retrieved
         /// using alGetError. Whenever alGetError is called, the error state is cleared and the last state (the current state
         /// when the call was made) is returned. To isolate error detection to a specific portion of code, alGetError should be
         /// called before the isolated section to clear the current error state.
         /// </returns>
-        [DllImport(Lib, EntryPoint = "alGetError", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGetError", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern ALError GetError();
-        // AL_API ALenum AL_APIENTRY alGetError( void );
 
         /// <summary>
         /// This function tests if a specific Extension is available for the OpenAL driver.
         /// </summary>
-        /// <param name="extname">A string naming the desired extension. Example: "EAX-RAM"</param>
+        /// <param name="extname">A string naming the desired extension. Example: "EAX-RAM".</param>
         /// <returns>Returns True if the Extension is available or False if not available.</returns>
-        [DllImport(Lib, EntryPoint = "alIsExtensionPresent", ExactSpelling = true, CallingConvention = Style,
-            CharSet = CharSet.Ansi)]
+        [DllImport(Library, EntryPoint = "alIsExtensionPresent", ExactSpelling = true, CallingConvention = Style, CharSet = CharSet.Ansi)]
         [SuppressUnmanagedCodeSecurity]
         public static extern bool IsExtensionPresent([In] string extname);
-        // AL_API ALboolean AL_APIENTRY alIsExtensionPresent( const ALchar* extname );
 
         /// <summary>
         /// This function returns the address of an OpenAL extension function. Handle with care.
@@ -212,81 +126,62 @@ namespace OpenTK.OpenAL.Native
         /// The return value is a pointer to the specified function. The return value will be IntPtr.Zero if the function
         /// is not found.
         /// </returns>
-        [DllImport(Lib, EntryPoint = "alGetProcAddress", ExactSpelling = true, CallingConvention = Style,
-            CharSet = CharSet.Ansi)]
+        [DllImport(Library, EntryPoint = "alGetProcAddress", ExactSpelling = true, CallingConvention = Style, CharSet = CharSet.Ansi)]
         [SuppressUnmanagedCodeSecurity]
         public static extern IntPtr GetProcAddress([In] string fname);
-        // AL_API void* AL_APIENTRY alGetProcAddress( const ALchar* fname );
 
         /// <summary>
         /// This function returns the enumeration value of an OpenAL token, described by a string.
         /// </summary>
-        /// <param name="ename">A string describing an OpenAL token. Example "AL_DISTANCE_MODEL"</param>
+        /// <param name="ename">A string describing an OpenAL token. Example "AL_DISTANCE_MODEL".</param>
         /// <returns>
         /// Returns the actual ALenum described by a string. Returns 0 if the string doesn’t describe a valid OpenAL
         /// token.
         /// </returns>
-        [DllImport(Lib, EntryPoint = "alGetEnumValue", ExactSpelling = true, CallingConvention = Style,
-            CharSet = CharSet.Ansi)]
+        [DllImport(Library, EntryPoint = "alGetEnumValue", ExactSpelling = true, CallingConvention = Style, CharSet = CharSet.Ansi)]
         [SuppressUnmanagedCodeSecurity]
         public static extern int GetEnumValue([In] string ename);
-        // AL_API ALenum AL_APIENTRY alGetEnumValue( const ALchar* ename );
-
-        /* Listener
-         * Listener represents the location and orientation of the
-         * 'user' in 3D-space.
-         *
-         * Properties include: -
-         *
-         * Gain         AL_GAIN         ALfloat
-         * Position     AL_POSITION     ALfloat[3]
-         * Velocity     AL_VELOCITY     ALfloat[3]
-         * Orientation  AL_ORIENTATION  ALfloat[6] (Forward then Up vectors)
-         */
 
         /// <summary>
         /// This function sets a floating-point property for the listener.
         /// </summary>
-        /// <param name="param">The name of the attribute to be set: ALListenerf.Gain</param>
+        /// <param name="param">The name of the attribute to be set: <see cref="ALListenerf.Gain"/>.</param>
         /// <param name="value">The float value to set the attribute to.</param>
-        [DllImport(Lib, EntryPoint = "alListenerf", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alListenerf", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void Listener(ALListenerf param, float value);
-        // AL_API void AL_APIENTRY alListenerf( ALenum param, ALfloat value );
 
         /// <summary>
         /// This function sets a floating-point property for the listener.
         /// </summary>
-        /// <param name="param">The name of the attribute to set: ALListener3f.Position, ALListener3f.Velocity</param>
-        /// <param name="value1">The value to set the attribute to.</param>
-        /// <param name="value2">The value to set the attribute to.</param>
-        /// <param name="value3">The value to set the attribute to.</param>
-        [DllImport(Lib, EntryPoint = "alListener3f", ExactSpelling = true, CallingConvention = Style)]
+        /// <param name="param">The name of the attribute to set: <see cref="ALListener3f.Position"/>, <see cref="ALListener3f.Velocity"/>.</param>
+        /// <param name="value1">The value to set the first element in the attribute to.</param>
+        /// <param name="value2">The value to set the second element in the attribute to.</param>
+        /// <param name="value3">The value to set the third element in the attribute to.</param>
+        [DllImport(Library, EntryPoint = "alListener3f", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void Listener(ALListener3f param, float value1, float value2, float value3);
-        // AL_API void AL_APIENTRY alListener3f( ALenum param, ALfloat value1, ALfloat value2, ALfloat value3 );
 
         /// <summary>
         /// This function sets a Math.Vector3 property for the listener.
         /// </summary>
-        /// <param name="param">The name of the attribute to set: ALListener3f.Position, ALListener3f.Velocity</param>
+        /// <param name="param">The name of the attribute to set: <see cref="ALListener3f.Position"/>, <see cref="ALListener3f.Velocity"/>.</param>
         /// <param name="values">The Math.Vector3 to set the attribute to.</param>
         public static void Listener(ALListener3f param, ref Vector3 values)
         {
             Listener(param, values.X, values.Y, values.Z);
         }
 
-        [DllImport(Lib, EntryPoint = "alListenerfv", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alListenerfv", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         private static extern unsafe void ListenerPrivate(ALListenerfv param, float* values);
-        // AL_API void AL_APIENTRY alListenerfv( ALenum param, const ALfloat* values );
 
         /// <summary>
         /// This function sets a floating-point vector property of the listener.
         /// </summary>
         /// <param name="param">
-        /// The name of the attribute to be set: ALListener3f.Position, ALListener3f.Velocity,
-        /// ALListenerfv.Orientation
+        /// The name of the attribute to be set: <see cref="ALListener3f.Position"/>, <see cref="ALListener3f.Velocity"/>,
+        /// <see cref="ALListenerfv.Orientation"/>.
         /// </param>
         /// <param name="values">Pointer to floating-point vector values.</param>
         public static void Listener(ALListenerfv param, ref float[] values)
@@ -303,7 +198,7 @@ namespace OpenTK.OpenAL.Native
         /// <summary>
         /// This function sets two Math.Vector3 properties of the listener.
         /// </summary>
-        /// <param name="param">The name of the attribute to be set: ALListenerfv.Orientation</param>
+        /// <param name="param">The name of the attribute to be set: <see cref="ALListenerfv.Orientation"/>.</param>
         /// <param name="at">A Math.Vector3 for the At-Vector.</param>
         /// <param name="up">A Math.Vector3 for the Up-Vector.</param>
         public static void Listener(ALListenerfv param, ref Vector3 at, ref Vector3 up)
@@ -327,38 +222,30 @@ namespace OpenTK.OpenAL.Native
             }
         }
 
-        // Not used by any Enums
-        // AL_API void AL_APIENTRY alListeneri( ALenum param, ALint value );
-        // AL_API void AL_APIENTRY alListener3i( ALenum param, ALint value1, ALint value2, ALint value3 );
-        // AL_API void AL_APIENTRY alListeneriv( ALenum param, const ALint* values );
-
         /// <summary>
         /// This function retrieves a floating-point property of the listener.
         /// </summary>
-        /// <param name="param">the name of the attribute to be retrieved: ALListenerf.Gain</param>
+        /// <param name="param">the name of the attribute to be retrieved: <see cref="ALListenerf.Gain"/>.</param>
         /// <param name="value">a pointer to the floating-point value being retrieved.</param>
-        [DllImport(Lib, EntryPoint = "alGetListenerf", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGetListenerf", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void GetListener(ALListenerf param, [Out] out float value);
-        // AL_API void AL_APIENTRY alGetListenerf( ALenum param, ALfloat* value );
 
         /// <summary>
         /// This function retrieves a set of three floating-point values from a property of the listener.
         /// </summary>
-        /// <param name="param">The name of the attribute to be retrieved: ALListener3f.Position, ALListener3f.Velocity</param>
+        /// <param name="param">The name of the attribute to be retrieved: <see cref="ALListener3f.Position"/>, <see cref="ALListener3f.Velocity"/>.</param>
         /// <param name="value1">The first floating-point value being retrieved.</param>
         /// <param name="value2">The second floating-point value  being retrieved.</param>
         /// <param name="value3">The third floating-point value  being retrieved.</param>
-        [DllImport(Lib, EntryPoint = "alGetListener3f", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGetListener3f", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
-        public static extern void GetListener(ALListener3f param, [Out] out float value1, [Out] out float value2,
-            [Out] out float value3);
-        // AL_API void AL_APIENTRY alGetListener3f( ALenum param, ALfloat *value1, ALfloat *value2, ALfloat *value3 );
+        public static extern void GetListener(ALListener3f param, [Out] out float value1, [Out] out float value2, [Out] out float value3);
 
         /// <summary>
         /// This function retrieves a Math.Vector3 from a property of the listener.
         /// </summary>
-        /// <param name="param">The name of the attribute to be retrieved: ALListener3f.Position, ALListener3f.Velocity</param>
+        /// <param name="param">The name of the attribute to be retrieved: <see cref="ALListener3f.Position"/>, <see cref="ALListener3f.Velocity"/>.</param>
         /// <param name="values">A Math.Vector3 to hold the three floats being retrieved.</param>
         public static void GetListener(ALListener3f param, out Vector3 values)
         {
@@ -369,19 +256,18 @@ namespace OpenTK.OpenAL.Native
         /// This function retrieves a floating-point vector property of the listener. You must pin it manually.
         /// </summary>
         /// <param name="param">
-        /// the name of the attribute to be retrieved: ALListener3f.Position, ALListener3f.Velocity,
-        /// ALListenerfv.Orientation
+        /// the name of the attribute to be retrieved: <see cref="ALListener3f.Position"/>, <see cref="ALListener3f.Velocity"/>,
+        /// <see cref="ALListenerfv.Orientation"/>.
         /// </param>
         /// <param name="values">A pointer to the floating-point vector value being retrieved.</param>
-        [DllImport(Lib, EntryPoint = "alGetListenerfv", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGetListenerfv", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void GetListener(ALListenerfv param, float* values);
-        // AL_API void AL_APIENTRY alGetListenerfv( ALenum param, ALfloat* values );
 
         /// <summary>
         /// This function retrieves two Math.Vector3 properties of the listener.
         /// </summary>
-        /// <param name="param">the name of the attribute to be retrieved: ALListenerfv.Orientation</param>
+        /// <param name="param">the name of the attribute to be retrieved: <see cref="ALListenerfv.Orientation"/>.</param>
         /// <param name="at">A Math.Vector3 for the At-Vector.</param>
         /// <param name="up">A Math.Vector3 for the Up-Vector.</param>
         public static void GetListener(ALListenerfv param, out Vector3 at, out Vector3 up)
@@ -404,56 +290,13 @@ namespace OpenTK.OpenAL.Native
             }
         }
 
-        // Not used by any Enum:
-        // AL_API void AL_APIENTRY alGetListeneri( ALenum param, ALint* value );
-        // AL_API void AL_APIENTRY alGetListener3i( ALenum param, ALint *value1, ALint *value2, ALint *value3 );
-        // AL_API void AL_APIENTRY alGetListeneriv( ALenum param, ALint* values );
-
-        /* Source
-         * Sources represent individual sound objects in 3D-space.
-         * Sources take the PCM buffer provided in the specified Buffer,
-         * apply Source-specific modifications, and then
-         * submit them to be mixed according to spatial arrangement etc.
-         *
-         * Properties include: -
-         *
-
-         * Position                          AL_POSITION             ALfloat[3]
-         * Velocity                          AL_VELOCITY             ALfloat[3]
-         * Direction                         AL_DIRECTION            ALfloat[3]
-
-         * Head Relative Mode                AL_SOURCE_RELATIVE      ALint (AL_TRUE or AL_FALSE)
-         * Looping                           AL_LOOPING              ALint (AL_TRUE or AL_FALSE)
-         *
-         * Reference Distance                AL_REFERENCE_DISTANCE   ALfloat
-         * Max Distance                      AL_MAX_DISTANCE         ALfloat
-         * RollOff Factor                    AL_ROLLOFF_FACTOR       ALfloat
-         * Pitch                             AL_PITCH                ALfloat
-         * Gain                              AL_GAIN                 ALfloat
-         * Min Gain                          AL_MIN_GAIN             ALfloat
-         * Max Gain                          AL_MAX_GAIN             ALfloat
-         * Inner Angle                       AL_CONE_INNER_ANGLE     ALint or ALfloat
-         * Outer Angle                       AL_CONE_OUTER_ANGLE     ALint or ALfloat
-         * Cone Outer Gain                   AL_CONE_OUTER_GAIN      ALint or ALfloat
-         *
-         * MS Offset                         AL_MSEC_OFFSET          ALint or ALfloat
-         * Byte Offset                       AL_BYTE_OFFSET          ALint or ALfloat
-         * Sample Offset                     AL_SAMPLE_OFFSET        ALint or ALfloat
-         * Attached Buffer                   AL_BUFFER               ALint
-         *
-         * State (Query only)                AL_SOURCE_STATE         ALint
-         * Buffers Queued (Query only)       AL_BUFFERS_QUEUED       ALint
-         * Buffers Processed (Query only)    AL_BUFFERS_PROCESSED    ALint
-         */
-
-        [DllImport(Lib, EntryPoint = "alGenSources", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGenSources", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         private static extern unsafe void GenSourcesPrivate(int n, [Out] uint* sources);
-        // AL_API void AL_APIENTRY alGenSources( ALsizei n, ALuint* Sources );
 
         /// <summary>
         /// This function generates one or more sources. References to sources are uint values, which are used wherever a
-        /// source reference is needed (in calls such as AL.DeleteSources and AL.Source with parameter ALSourcei).
+        /// source reference is needed (in calls such as <see cref="AL.DeleteSources"/> and <see cref="AL.Source"/> with parameter ALSourcei).
         /// </summary>
         /// <param name="n">The number of sources to be generated.</param>
         /// <param name="sources">Pointer to an array of uint values which will store the names of the new sources.</param>
@@ -470,7 +313,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function generates one or more sources. References to sources are int values, which are used wherever a
-        /// source reference is needed (in calls such as AL.DeleteSources and AL.Source with parameter ALSourcei).
+        /// source reference is needed (in calls such as <see cref="AL.DeleteSources"/> and <see cref="AL.Source"/> with parameter ALSourcei).
         /// </summary>
         /// <param name="n">The number of sources to be generated.</param>
         /// <param name="sources">Pointer to an array of int values which will store the names of the new sources.</param>
@@ -487,7 +330,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function generates one or more sources. References to sources are int values, which are used wherever a
-        /// source reference is needed (in calls such as AL.DeleteSources and AL.Source with parameter ALSourcei).
+        /// source reference is needed (in calls such as <see cref="AL.DeleteSources"/> and <see cref="AL.Source"/> with parameter ALSourcei).
         /// </summary>
         /// <param name="sources">Pointer to an array of int values which will store the names of the new sources.</param>
         public static void GenSources(int[] sources)
@@ -502,7 +345,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function generates one or more sources. References to sources are int values, which are used wherever a
-        /// source reference is needed (in calls such as AL.DeleteSources and AL.Source with parameter ALSourcei).
+        /// source reference is needed (in calls such as <see cref="AL.DeleteSources"/> and <see cref="AL.Source"/> with parameter ALSourcei).
         /// </summary>
         /// <param name="n">The number of sources to be generated.</param>
         /// <returns>Pointer to an array of int values which will store the names of the new sources.</returns>
@@ -521,7 +364,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function generates one source only. References to sources are int values, which are used wherever a
-        /// source reference is needed (in calls such as AL.DeleteSources and AL.Source with parameter ALSourcei).
+        /// source reference is needed (in calls such as <see cref="AL.DeleteSources"/> and <see cref="AL.Source"/> with parameter ALSourcei).
         /// </summary>
         /// <returns>Pointer to an int value which will store the name of the new source.</returns>
         public static int GenSource()
@@ -532,7 +375,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function generates one source only. References to sources are uint values, which are used wherever a
-        /// source reference is needed (in calls such as AL.DeleteSources and AL.Source with parameter ALSourcei).
+        /// source reference is needed (in calls such as <see cref="AL.DeleteSources"/> and <see cref="AL.Source"/> with parameter ALSourcei).
         /// </summary>
         /// <param name="source">Pointer to an uint value which will store the name of the new source.</param>
         public static void GenSource(out uint source)
@@ -545,17 +388,16 @@ namespace OpenTK.OpenAL.Native
         /// </summary>
         /// <param name="n">The number of sources to be deleted.</param>
         /// <param name="sources">Pointer to an array of source names identifying the sources to be deleted.</param>
-        [DllImport(Lib, EntryPoint = "alDeleteSources", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alDeleteSources", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
-        public static extern unsafe void DeleteSources(int n, [In] uint* sources); // Delete Source objects
-        // AL_API void AL_APIENTRY alDeleteSources( ALsizei n, const ALuint* Sources );
+        public static extern unsafe void DeleteSources(int n, [In] uint* sources);
 
         /// <summary>
         /// This function deletes one or more sources.
         /// </summary>
         /// <param name="n">The number of sources to be deleted.</param>
         /// <param name="sources">Reference to a single source, or an array of source names identifying the sources to be deleted.</param>
-        [DllImport(Lib, EntryPoint = "alDeleteSources", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alDeleteSources", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void DeleteSources(int n, ref uint sources);
 
@@ -564,7 +406,7 @@ namespace OpenTK.OpenAL.Native
         /// </summary>
         /// <param name="n">The number of sources to be deleted.</param>
         /// <param name="sources">Reference to a single source, or an array of source names identifying the sources to be deleted.</param>
-        [DllImport(Lib, EntryPoint = "alDeleteSources", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alDeleteSources", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void DeleteSources(int n, ref int sources);
 
@@ -627,17 +469,16 @@ namespace OpenTK.OpenAL.Native
         /// <summary>
         /// This function tests if a source name is valid, returning True if valid and False if not.
         /// </summary>
-        /// <param name="sid">A source name to be tested for validity</param>
+        /// <param name="sid">A source name to be tested for validity.</param>
         /// <returns>Success.</returns>
-        [DllImport(Lib, EntryPoint = "alIsSource", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alIsSource", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern bool IsSource(uint sid);
-        // AL_API ALboolean AL_APIENTRY alIsSource( ALuint sid );
 
         /// <summary>
         /// This function tests if a source name is valid, returning True if valid and False if not.
         /// </summary>
-        /// <param name="sid">A source name to be tested for validity</param>
+        /// <param name="sid">A source name to be tested for validity.</param>
         /// <returns>Success.</returns>
         public static bool IsSource(int sid)
         {
@@ -647,24 +488,23 @@ namespace OpenTK.OpenAL.Native
         /// <summary>
         /// This function sets a floating-point property of a source.
         /// </summary>
-        /// <param name="sid">Source name whose attribute is being set</param>
+        /// <param name="sid">Source name whose attribute is being set.</param>
         /// <param name="param">
-        /// The name of the attribute to set: ALSourcef.Pitch, Gain, MinGain, MaxGain, MaxDistance,
+        /// The name of the attribute to set: <see cref="ALSourcef.Pitch"/>, Gain, MinGain, MaxGain, MaxDistance,
         /// RolloffFactor, ConeOuterGain, ConeInnerAngle, ConeOuterAngle, ReferenceDistance, EfxAirAbsorptionFactor,
         /// EfxRoomRolloffFactor, EfxConeOuterGainHighFrequency.
         /// </param>
         /// <param name="value">The value to set the attribute to.</param>
-        [DllImport(Lib, EntryPoint = "alSourcef", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alSourcef", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void Source(uint sid, ALSourcef param, float value);
-        // AL_API void AL_APIENTRY alSourcef( ALuint sid, ALenum param, ALfloat value );
 
         /// <summary>
         /// This function sets a floating-point property of a source.
         /// </summary>
-        /// <param name="sid">Source name whose attribute is being set</param>
+        /// <param name="sid">Source name whose attribute is being set.</param>
         /// <param name="param">
-        /// The name of the attribute to set: ALSourcef.Pitch, Gain, MinGain, MaxGain, MaxDistance,
+        /// The name of the attribute to set: <see cref="ALSourcef.Pitch"/>, Gain, MinGain, MaxGain, MaxDistance,
         /// RolloffFactor, ConeOuterGain, ConeInnerAngle, ConeOuterAngle, ReferenceDistance, EfxAirAbsorptionFactor,
         /// EfxRoomRolloffFactor, EfxConeOuterGainHighFrequency.
         /// </param>
@@ -678,23 +518,22 @@ namespace OpenTK.OpenAL.Native
         /// This function sets a source property requiring three floating-point values.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
-        /// <param name="param">The name of the attribute to set: ALSource3f.Position, Velocity, Direction.</param>
-        /// <param name="value1">The three ALfloat values which the attribute will be set to.</param>
-        /// <param name="value2">The three ALfloat values which the attribute will be set to.</param>
-        /// <param name="value3">The three ALfloat values which the attribute will be set to.</param>
-        [DllImport(Lib, EntryPoint = "alSource3f", ExactSpelling = true, CallingConvention = Style)]
+        /// <param name="param">The name of the attribute to set: <see cref="ALSource3f.Position"/>, Velocity, Direction.</param>
+        /// <param name="value1">The first ALfloat value which the attribute will be set to.</param>
+        /// <param name="value2">The second ALfloat value which the attribute will be set to.</param>
+        /// <param name="value3">The third ALfloat value which the attribute will be set to.</param>
+        [DllImport(Library, EntryPoint = "alSource3f", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void Source(uint sid, ALSource3f param, float value1, float value2, float value3);
-        // AL_API void AL_APIENTRY alSource3f( ALuint sid, ALenum param, ALfloat value1, ALfloat value2, ALfloat value3 );
 
         /// <summary>
         /// This function sets a source property requiring three floating-point values.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
-        /// <param name="param">The name of the attribute to set: ALSource3f.Position, Velocity, Direction.</param>
-        /// <param name="value1">The three ALfloat values which the attribute will be set to.</param>
-        /// <param name="value2">The three ALfloat values which the attribute will be set to.</param>
-        /// <param name="value3">The three ALfloat values which the attribute will be set to.</param>
+        /// <param name="param">The name of the attribute to set: <see cref="ALSource3f.Position"/>, Velocity, Direction.</param>
+        /// <param name="value1">The first ALfloat value which the attribute will be set to.</param>
+        /// <param name="value2">The second ALfloat value which the attribute will be set to.</param>
+        /// <param name="value3">The third ALfloat value which the attribute will be set to.</param>
         public static void Source(int sid, ALSource3f param, float value1, float value2, float value3)
         {
             Source((uint)sid, param, value1, value2, value3);
@@ -704,7 +543,7 @@ namespace OpenTK.OpenAL.Native
         /// This function sets a source property requiring three floating-point values.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
-        /// <param name="param">The name of the attribute to set: ALSource3f.Position, Velocity, Direction.</param>
+        /// <param name="param">The name of the attribute to set: <see cref="ALSource3f.Position"/>, Velocity, Direction.</param>
         /// <param name="values">A Math.Vector3 which the attribute will be set to.</param>
         public static void Source(uint sid, ALSource3f param, ref Vector3 values)
         {
@@ -715,7 +554,7 @@ namespace OpenTK.OpenAL.Native
         /// This function sets a source property requiring three floating-point values.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
-        /// <param name="param">The name of the attribute to set: ALSource3f.Position, Velocity, Direction.</param>
+        /// <param name="param">The name of the attribute to set: <see cref="ALSource3f.Position"/>, Velocity, Direction.</param>
         /// <param name="values">A Math.Vector3 which the attribute will be set to.</param>
         public static void Source(int sid, ALSource3f param, ref Vector3 values)
         {
@@ -727,21 +566,20 @@ namespace OpenTK.OpenAL.Native
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
         /// <param name="param">
-        /// The name of the attribute to set: ALSourcei.SourceRelative, ConeInnerAngle, ConeOuterAngle,
+        /// The name of the attribute to set: <see cref="ALSourcei.SourceRelative"/>, ConeInnerAngle, ConeOuterAngle,
         /// Looping, Buffer, SourceState.
         /// </param>
         /// <param name="value">The value to set the attribute to.</param>
-        [DllImport(Lib, EntryPoint = "alSourcei", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alSourcei", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void Source(uint sid, ALSourcei param, int value);
-        // AL_API void AL_APIENTRY alSourcei( ALuint sid, ALenum param, ALint value );
 
         /// <summary>
         /// This function sets an integer property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
         /// <param name="param">
-        /// The name of the attribute to set: ALSourcei.SourceRelative, ConeInnerAngle, ConeOuterAngle,
+        /// The name of the attribute to set: <see cref="ALSourcei.SourceRelative"/>, ConeInnerAngle, ConeOuterAngle,
         /// Looping, Buffer, SourceState.
         /// </param>
         /// <param name="value">The value to set the attribute to.</param>
@@ -754,7 +592,7 @@ namespace OpenTK.OpenAL.Native
         /// This function sets an bool property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
-        /// <param name="param">The name of the attribute to set: ALSourceb.SourceRelative, Looping.</param>
+        /// <param name="param">The name of the attribute to set: <see cref="ALSourceb.SourceRelative"/>, Looping.</param>
         /// <param name="value">The value to set the attribute to.</param>
         public static void Source(uint sid, ALSourceb param, bool value)
         {
@@ -765,7 +603,7 @@ namespace OpenTK.OpenAL.Native
         /// This function sets an bool property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
-        /// <param name="param">The name of the attribute to set: ALSourceb.SourceRelative, Looping.</param>
+        /// <param name="param">The name of the attribute to set: <see cref="ALSourceb.SourceRelative"/>, Looping.</param>
         /// <param name="value">The value to set the attribute to.</param>
         public static void Source(int sid, ALSourceb param, bool value)
         {
@@ -797,22 +635,21 @@ namespace OpenTK.OpenAL.Native
         /// Sources and Auxiliary Effect Slots.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
-        /// <param name="param">The name of the attribute to set: EfxAuxiliarySendFilter</param>
-        /// <param name="value1">The value to set the attribute to. (EFX Extension) The destination Auxiliary Effect Slot ID</param>
+        /// <param name="param">The name of the attribute to set: EfxAuxiliarySendFilter.</param>
+        /// <param name="value1">The value to set the attribute to. (EFX Extension) The destination Auxiliary Effect Slot ID.</param>
         /// <param name="value2">The value to set the attribute to. (EFX Extension) The Auxiliary Send number.</param>
         /// <param name="value3">The value to set the attribute to. (EFX Extension) optional Filter ID.</param>
-        [DllImport(Lib, EntryPoint = "alSource3i", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alSource3i", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void Source(uint sid, ALSource3i param, int value1, int value2, int value3);
-        // AL_API void AL_APIENTRY alSource3i( ALuint sid, ALenum param, ALint value1, ALint value2, ALint value3 );
 
         /// <summary>
         /// This function sets 3 integer properties of a source. This property is used to establish connections between
         /// Sources and Auxiliary Effect Slots.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being set.</param>
-        /// <param name="param">The name of the attribute to set: EfxAuxiliarySendFilter</param>
-        /// <param name="value1">The value to set the attribute to. (EFX Extension) The destination Auxiliary Effect Slot ID</param>
+        /// <param name="param">The name of the attribute to set: EfxAuxiliarySendFilter.</param>
+        /// <param name="value1">The value to set the attribute to. (EFX Extension) The destination Auxiliary Effect Slot ID.</param>
         /// <param name="value2">The value to set the attribute to. (EFX Extension) The Auxiliary Send number.</param>
         /// <param name="value3">The value to set the attribute to. (EFX Extension) optional Filter ID.</param>
         public static void Source(int sid, ALSource3i param, int value1, int value2, int value3)
@@ -820,35 +657,30 @@ namespace OpenTK.OpenAL.Native
             Source((uint)sid, param, value1, value2, value3);
         }
 
-        // Not used by any Enum:
-        // AL_API void AL_APIENTRY alSourcefv( ALuint sid, ALenum param, const ALfloat* values );
-        // AL_API void AL_APIENTRY alSourceiv( ALuint sid, ALenum param, const ALint* values );
-
         /// <summary>
         /// This function retrieves a floating-point property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
         /// <param name="param">
-        /// The name of the attribute to set: ALSourcef.Pitch, Gain, MinGain, MaxGain, MaxDistance,
+        /// The name of the attribute to set: <see cref="ALSourcef.Pitch"/>, Gain, MinGain, MaxGain, MaxDistance,
         /// RolloffFactor, ConeOuterGain, ConeInnerAngle, ConeOuterAngle, ReferenceDistance, EfxAirAbsorptionFactor,
         /// EfxRoomRolloffFactor, EfxConeOuterGainHighFrequency.
         /// </param>
-        /// <param name="value">A pointer to the floating-point value being retrieved</param>
-        [DllImport(Lib, EntryPoint = "alGetSourcef", ExactSpelling = true, CallingConvention = Style)]
+        /// <param name="value">A pointer to the floating-point value being retrieved.</param>
+        [DllImport(Library, EntryPoint = "alGetSourcef", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void GetSource(uint sid, ALSourcef param, [Out] out float value);
-        // AL_API void AL_APIENTRY alGetSourcef( ALuint sid, ALenum param, ALfloat* value );
 
         /// <summary>
         /// This function retrieves a floating-point property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
         /// <param name="param">
-        /// The name of the attribute to set: ALSourcef.Pitch, Gain, MinGain, MaxGain, MaxDistance,
+        /// The name of the attribute to set: <see cref="ALSourcef.Pitch"/>, Gain, MinGain, MaxGain, MaxDistance,
         /// RolloffFactor, ConeOuterGain, ConeInnerAngle, ConeOuterAngle, ReferenceDistance, EfxAirAbsorptionFactor,
         /// EfxRoomRolloffFactor, EfxConeOuterGainHighFrequency.
         /// </param>
-        /// <param name="value">A pointer to the floating-point value being retrieved</param>
+        /// <param name="value">A pointer to the floating-point value being retrieved.</param>
         public static void GetSource(int sid, ALSourcef param, out float value)
         {
             GetSource((uint)sid, param, out value);
@@ -858,24 +690,22 @@ namespace OpenTK.OpenAL.Native
         /// This function retrieves three floating-point values representing a property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
-        /// <param name="param">the name of the attribute being retrieved: ALSource3f.Position, Velocity, Direction.</param>
-        /// <param name="value1">Pointer to the value to retrieve.</param>
-        /// <param name="value2">Pointer to the value to retrieve.</param>
-        /// <param name="value3">Pointer to the value to retrieve.</param>
-        [DllImport(Lib, EntryPoint = "alGetSource3f", ExactSpelling = true, CallingConvention = Style)]
+        /// <param name="param">the name of the attribute being retrieved: <see cref="ALSource3f.Position"/>, Velocity, Direction.</param>
+        /// <param name="value1">Pointer to the first element in the  value to retrieve.</param>
+        /// <param name="value2">Pointer to the second element in the  value to retrieve.</param>
+        /// <param name="value3">Pointer to the third element in the  value to retrieve.</param>
+        [DllImport(Library, EntryPoint = "alGetSource3f", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
-        public static extern void GetSource(uint sid, ALSource3f param, [Out] out float value1, [Out] out float value2,
-            [Out] out float value3);
-        // AL_API void AL_APIENTRY alGetSource3f( ALuint sid, ALenum param, ALfloat* value1, ALfloat* value2, ALfloat* value3);
+        public static extern void GetSource(uint sid, ALSource3f param, [Out] out float value1, [Out] out float value2, [Out] out float value3);
 
         /// <summary>
         /// This function retrieves three floating-point values representing a property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
-        /// <param name="param">the name of the attribute being retrieved: ALSource3f.Position, Velocity, Direction.</param>
-        /// <param name="value1">Pointer to the value to retrieve.</param>
-        /// <param name="value2">Pointer to the value to retrieve.</param>
-        /// <param name="value3">Pointer to the value to retrieve.</param>
+        /// <param name="param">the name of the attribute being retrieved: <see cref="ALSource3f.Position"/>, Velocity, Direction.</param>
+        /// <param name="value1">Pointer to the first element in the  value to retrieve.</param>
+        /// <param name="value2">Pointer to the second element in the  value to retrieve.</param>
+        /// <param name="value3">Pointer to the third element in the  value to retrieve.</param>
         public static void GetSource(int sid, ALSource3f param, out float value1, out float value2, out float value3)
         {
             GetSource((uint)sid, param, out value1, out value2, out value3);
@@ -885,7 +715,7 @@ namespace OpenTK.OpenAL.Native
         /// This function retrieves three floating-point values representing a property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
-        /// <param name="param">the name of the attribute being retrieved: ALSource3f.Position, Velocity, Direction.</param>
+        /// <param name="param">the name of the attribute being retrieved: <see cref="ALSource3f.Position"/>, Velocity, Direction.</param>
         /// <param name="values">A Math.Vector3 to retrieve the values to.</param>
         public static void GetSource(uint sid, ALSource3f param, out Vector3 values)
         {
@@ -896,7 +726,7 @@ namespace OpenTK.OpenAL.Native
         /// This function retrieves three floating-point values representing a property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
-        /// <param name="param">the name of the attribute being retrieved: ALSource3f.Position, Velocity, Direction.</param>
+        /// <param name="param">the name of the attribute being retrieved: <see cref="ALSource3f.Position"/>, Velocity, Direction.</param>
         /// <param name="values">A Math.Vector3 to retrieve the values to.</param>
         public static void GetSource(int sid, ALSource3f param, out Vector3 values)
         {
@@ -908,21 +738,20 @@ namespace OpenTK.OpenAL.Native
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
         /// <param name="param">
-        /// The name of the attribute to retrieve: ALSourcei.SourceRelative, Buffer, SourceState,
+        /// The name of the attribute to retrieve: <see cref="ALSourcei.SourceRelative"/>, Buffer, SourceState,
         /// BuffersQueued, BuffersProcessed.
         /// </param>
         /// <param name="value">A pointer to the integer value being retrieved.</param>
-        [DllImport(Lib, EntryPoint = "alGetSourcei", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGetSourcei", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void GetSource(uint sid, ALGetSourcei param, [Out] out int value);
-        // AL_API void AL_APIENTRY alGetSourcei( ALuint sid,  ALenum param, ALint* value );
 
         /// <summary>
         /// This function retrieves an integer property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
         /// <param name="param">
-        /// The name of the attribute to retrieve: ALSourcei.SourceRelative, Buffer, SourceState,
+        /// The name of the attribute to retrieve: <see cref="ALSourcei.SourceRelative"/>, Buffer, SourceState,
         /// BuffersQueued, BuffersProcessed.
         /// </param>
         /// <param name="value">A pointer to the integer value being retrieved.</param>
@@ -935,7 +764,7 @@ namespace OpenTK.OpenAL.Native
         /// This function retrieves a bool property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
-        /// <param name="param">The name of the attribute to get: ALSourceb.SourceRelative, Looping.</param>
+        /// <param name="param">The name of the attribute to get: <see cref="ALSourceb.SourceRelative"/>, Looping.</param>
         /// <param name="value">A pointer to the bool value being retrieved.</param>
         public static void GetSource(uint sid, ALSourceb param, out bool value)
         {
@@ -947,7 +776,7 @@ namespace OpenTK.OpenAL.Native
         /// This function retrieves a bool property of a source.
         /// </summary>
         /// <param name="sid">Source name whose attribute is being retrieved.</param>
-        /// <param name="param">The name of the attribute to get: ALSourceb.SourceRelative, Looping.</param>
+        /// <param name="param">The name of the attribute to get: <see cref="ALSourceb.SourceRelative"/>, Looping.</param>
         /// <param name="value">A pointer to the bool value being retrieved.</param>
         public static void GetSource(int sid, ALSourceb param, out bool value)
         {
@@ -955,27 +784,21 @@ namespace OpenTK.OpenAL.Native
             value = result != 0;
         }
 
-        // Not used by any Enum:
-        // AL_API void AL_APIENTRY alGetSource3i( ALuint sid, ALenum param, ALint* value1, ALint* value2, ALint* value3);
-        // AL_API void AL_APIENTRY alGetSourcefv( ALuint sid, ALenum param, ALfloat* values );
-        // AL_API void AL_APIENTRY alGetSourceiv( ALuint sid,  ALenum param, ALint* values );
-
         /// <summary>
         /// This function plays a set of sources. The playing sources will have their state changed to
-        /// ALSourceState.Playing. When called on a source which is already playing, the source will restart at the beginning.
-        /// When the attached buffer(s) are done playing, the source will progress to the ALSourceState.Stopped state.
+        /// <see cref="ALSourceState.Playing"/>. When called on a source which is already playing, the source will restart at the beginning.
+        /// When the attached buffer(s) are done playing, the source will progress to the <see cref="ALSourceState.Stopped"/> state.
         /// </summary>
         /// <param name="ns">The number of sources to be played.</param>
         /// <param name="sids">A pointer to an array of sources to be played.</param>
-        [DllImport(Lib, EntryPoint = "alSourcePlayv")]
+        [DllImport(Library, EntryPoint = "alSourcePlayv")]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void SourcePlay(int ns, [In] uint* sids);
-        // AL_API void AL_APIENTRY alSourcePlayv( ALsizei ns, const ALuint *sids );
 
         /// <summary>
         /// This function plays a set of sources. The playing sources will have their state changed to
-        /// ALSourceState.Playing. When called on a source which is already playing, the source will restart at the beginning.
-        /// When the attached buffer(s) are done playing, the source will progress to the ALSourceState.Stopped state.
+        /// <see cref="ALSourceState.Playing"/>. When called on a source which is already playing, the source will restart at the beginning.
+        /// When the attached buffer(s) are done playing, the source will progress to the <see cref="ALSourceState.Stopped"/> state.
         /// </summary>
         /// <param name="ns">The number of sources to be played.</param>
         /// <param name="sids">A pointer to an array of sources to be played.</param>
@@ -992,8 +815,8 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function plays a set of sources. The playing sources will have their state changed to
-        /// ALSourceState.Playing. When called on a source which is already playing, the source will restart at the beginning.
-        /// When the attached buffer(s) are done playing, the source will progress to the ALSourceState.Stopped state.
+        /// <see cref="ALSourceState.Playing"/>. When called on a source which is already playing, the source will restart at the beginning.
+        /// When the attached buffer(s) are done playing, the source will progress to the <see cref="ALSourceState.Stopped"/> state.
         /// </summary>
         /// <param name="ns">The number of sources to be played.</param>
         /// <param name="sids">A pointer to an array of sources to be played.</param>
@@ -1010,8 +833,8 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function plays a set of sources. The playing sources will have their state changed to
-        /// ALSourceState.Playing. When called on a source which is already playing, the source will restart at the beginning.
-        /// When the attached buffer(s) are done playing, the source will progress to the ALSourceState.Stopped state.
+        /// <see cref="ALSourceState.Playing"/>. When called on a source which is already playing, the source will restart at the beginning.
+        /// When the attached buffer(s) are done playing, the source will progress to the <see cref="ALSourceState.Stopped"/> state.
         /// </summary>
         /// <param name="ns">The number of sources to be played.</param>
         /// <param name="sids">A pointer to an array of sources to be played.</param>
@@ -1028,18 +851,17 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function stops a set of sources. The stopped sources will have their state changed to
-        /// ALSourceState.Stopped.
+        /// <see cref="ALSourceState.Stopped"/>.
         /// </summary>
         /// <param name="ns">The number of sources to stop.</param>
         /// <param name="sids">A pointer to an array of sources to be stopped.</param>
-        [DllImport(Lib, EntryPoint = "alSourceStopv")]
+        [DllImport(Library, EntryPoint = "alSourceStopv")]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void SourceStop(int ns, [In] uint* sids);
-        // AL_API void AL_APIENTRY alSourceStopv( ALsizei ns, const ALuint *sids );
 
         /// <summary>
         /// This function stops a set of sources. The stopped sources will have their state changed to
-        /// ALSourceState.Stopped.
+        /// <see cref="ALSourceState.Stopped"/>.
         /// </summary>
         /// <param name="ns">The number of sources to stop.</param>
         /// <param name="sids">A pointer to an array of sources to be stopped.</param>
@@ -1056,7 +878,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function stops a set of sources. The stopped sources will have their state changed to
-        /// ALSourceState.Stopped.
+        /// <see cref="ALSourceState.Stopped"/>.
         /// </summary>
         /// <param name="ns">The number of sources to stop.</param>
         /// <param name="sids">A pointer to an array of sources to be stopped.</param>
@@ -1073,7 +895,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function stops a set of sources. The stopped sources will have their state changed to
-        /// ALSourceState.Stopped.
+        /// <see cref="ALSourceState.Stopped"/>.
         /// </summary>
         /// <param name="ns">The number of sources to stop.</param>
         /// <param name="sids">A pointer to an array of sources to be stopped.</param>
@@ -1089,17 +911,16 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function stops a set of sources and sets all their states to ALSourceState.Initial.
+        /// This function stops a set of sources and sets all their states to <see cref="ALSourceState.Initial"/>.
         /// </summary>
         /// <param name="ns">The number of sources to be rewound.</param>
         /// <param name="sids">A pointer to an array of sources to be rewound.</param>
-        [DllImport(Lib, EntryPoint = "alSourceRewindv")]
+        [DllImport(Library, EntryPoint = "alSourceRewindv")]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void SourceRewind(int ns, [In] uint* sids);
-        // AL_API void AL_APIENTRY alSourceRewindv( ALsizei ns, const ALuint *sids );
 
         /// <summary>
-        /// This function stops a set of sources and sets all their states to ALSourceState.Initial.
+        /// This function stops a set of sources and sets all their states to <see cref="ALSourceState.Initial"/>.
         /// </summary>
         /// <param name="ns">The number of sources to be rewound.</param>
         /// <param name="sids">A pointer to an array of sources to be rewound.</param>
@@ -1115,7 +936,7 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function stops a set of sources and sets all their states to ALSourceState.Initial.
+        /// This function stops a set of sources and sets all their states to <see cref="ALSourceState.Initial"/>.
         /// </summary>
         /// <param name="ns">The number of sources to be rewound.</param>
         /// <param name="sids">A pointer to an array of sources to be rewound.</param>
@@ -1131,7 +952,7 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function stops a set of sources and sets all their states to ALSourceState.Initial.
+        /// This function stops a set of sources and sets all their states to <see cref="ALSourceState.Initial"/>.
         /// </summary>
         /// <param name="ns">The number of sources to be rewound.</param>
         /// <param name="sids">A pointer to an array of sources to be rewound.</param>
@@ -1148,18 +969,17 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function pauses a set of sources. The paused sources will have their state changed to
-        /// ALSourceState.Paused.
+        /// <see cref="ALSourceState.Paused"/>.
         /// </summary>
         /// <param name="ns">The number of sources to be paused.</param>
         /// <param name="sids">A pointer to an array of sources to be paused.</param>
-        [DllImport(Lib, EntryPoint = "alSourcePausev")]
+        [DllImport(Library, EntryPoint = "alSourcePausev")]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void SourcePause(int ns, [In] uint* sids);
-        // AL_API void AL_APIENTRY alSourcePausev( ALsizei ns, const ALuint *sids );
 
         /// <summary>
         /// This function pauses a set of sources. The paused sources will have their state changed to
-        /// ALSourceState.Paused.
+        /// <see cref="ALSourceState.Paused"/>.
         /// </summary>
         /// <param name="ns">The number of sources to be paused.</param>
         /// <param name="sids">A pointer to an array of sources to be paused.</param>
@@ -1176,7 +996,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function pauses a set of sources. The paused sources will have their state changed to
-        /// ALSourceState.Paused.
+        /// <see cref="ALSourceState.Paused"/>.
         /// </summary>
         /// <param name="ns">The number of sources to be paused.</param>
         /// <param name="sids">A pointer to an array of sources to be paused.</param>
@@ -1193,7 +1013,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function pauses a set of sources. The paused sources will have their state changed to
-        /// ALSourceState.Paused.
+        /// <see cref="ALSourceState.Paused"/>.
         /// </summary>
         /// <param name="ns">The number of sources to be paused.</param>
         /// <param name="sids">A pointer to an array of sources to be paused.</param>
@@ -1210,19 +1030,18 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function plays, replays or resumes a source. The playing source will have it's state changed to
-        /// ALSourceState.Playing. When called on a source which is already playing, the source will restart at the beginning.
-        /// When the attached buffer(s) are done playing, the source will progress to the ALSourceState.Stopped state.
+        /// <see cref="ALSourceState.Playing"/>. When called on a source which is already playing, the source will restart at the beginning.
+        /// When the attached buffer(s) are done playing, the source will progress to the <see cref="ALSourceState.Stopped"/> state.
         /// </summary>
         /// <param name="sid">The name of the source to be played.</param>
-        [DllImport(Lib, EntryPoint = "alSourcePlay", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alSourcePlay", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void SourcePlay(uint sid);
-        // AL_API void AL_APIENTRY alSourcePlay( ALuint sid );
 
         /// <summary>
         /// This function plays, replays or resumes a source. The playing source will have it's state changed to
-        /// ALSourceState.Playing. When called on a source which is already playing, the source will restart at the beginning.
-        /// When the attached buffer(s) are done playing, the source will progress to the ALSourceState.Stopped state.
+        /// <see cref="ALSourceState.Playing"/>. When called on a source which is already playing, the source will restart at the beginning.
+        /// When the attached buffer(s) are done playing, the source will progress to the <see cref="ALSourceState.Stopped"/> state.
         /// </summary>
         /// <param name="sid">The name of the source to be played.</param>
         public static void SourcePlay(int sid)
@@ -1231,16 +1050,15 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function stops a source. The stopped source will have it's state changed to ALSourceState.Stopped.
+        /// This function stops a source. The stopped source will have it's state changed to <see cref="ALSourceState.Stopped"/>.
         /// </summary>
         /// <param name="sid">The name of the source to be stopped.</param>
-        [DllImport(Lib, EntryPoint = "alSourceStop", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alSourceStop", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void SourceStop(uint sid);
-        // AL_API void AL_APIENTRY alSourceStop( ALuint sid );
 
         /// <summary>
-        /// This function stops a source. The stopped source will have it's state changed to ALSourceState.Stopped.
+        /// This function stops a source. The stopped source will have it's state changed to <see cref="ALSourceState.Stopped"/>.
         /// </summary>
         /// <param name="sid">The name of the source to be stopped.</param>
         public static void SourceStop(int sid)
@@ -1249,16 +1067,15 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function stops the source and sets its state to ALSourceState.Initial.
+        /// This function stops the source and sets its state to <see cref="ALSourceState.Initial"/>.
         /// </summary>
         /// <param name="sid">The name of the source to be rewound.</param>
-        [DllImport(Lib, EntryPoint = "alSourceRewind", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alSourceRewind", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void SourceRewind(uint sid);
-        // AL_API void AL_APIENTRY alSourceRewind( ALuint sid );
 
         /// <summary>
-        /// This function stops the source and sets its state to ALSourceState.Initial.
+        /// This function stops the source and sets its state to <see cref="ALSourceState.Initial"/>.
         /// </summary>
         /// <param name="sid">The name of the source to be rewound.</param>
         public static void SourceRewind(int sid)
@@ -1267,16 +1084,15 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function pauses a source. The paused source will have its state changed to ALSourceState.Paused.
+        /// This function pauses a source. The paused source will have its state changed to <see cref="ALSourceState.Paused"/>.
         /// </summary>
         /// <param name="sid">The name of the source to be paused.</param>
-        [DllImport(Lib, EntryPoint = "alSourcePause", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alSourcePause", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void SourcePause(uint sid);
-        // AL_API void AL_APIENTRY alSourcePause( ALuint sid );
 
         /// <summary>
-        /// This function pauses a source. The paused source will have its state changed to ALSourceState.Paused.
+        /// This function pauses a source. The paused source will have its state changed to <see cref="ALSourceState.Paused"/>.
         /// </summary>
         /// <param name="sid">The name of the source to be paused.</param>
         public static void SourcePause(int sid)
@@ -1286,23 +1102,22 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function queues a set of buffers on a source. All buffers attached to a source will be played in
-        /// sequence, and the number of processed buffers can be detected using AL.GetSource with parameter
-        /// ALGetSourcei.BuffersProcessed. When first created, a source will be of type ALSourceType.Undetermined. A successful
-        /// AL.SourceQueueBuffers call will change the source type to ALSourceType.Streaming.
+        /// sequence, and the number of processed buffers can be detected using <see cref="AL.GetSource"/> with parameter
+        /// <see cref="ALGetSourcei.BuffersProcessed"/>. When first created, a source will be of type <see cref="ALSourceType.Undetermined"/>. A successful
+        /// <see cref="AL.SourceQueueBuffers"/> call will change the source type to <see cref="ALSourceType.Streaming"/>.
         /// </summary>
         /// <param name="sid">The name of the source to queue buffers onto.</param>
         /// <param name="numEntries">The number of buffers to be queued.</param>
         /// <param name="bids">A pointer to an array of buffer names to be queued.</param>
-        [DllImport(Lib, EntryPoint = "alSourceQueueBuffers")]
+        [DllImport(Library, EntryPoint = "alSourceQueueBuffers")]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void SourceQueueBuffers(uint sid, int numEntries, [In] uint* bids);
-        // AL_API void AL_APIENTRY alSourceQueueBuffers( ALuint sid, ALsizei numEntries, const ALuint *bids );
 
         /// <summary>
         /// This function queues a set of buffers on a source. All buffers attached to a source will be played in
-        /// sequence, and the number of processed buffers can be detected using AL.GetSource with parameter
-        /// ALGetSourcei.BuffersProcessed. When first created, a source will be of type ALSourceType.Undetermined. A successful
-        /// AL.SourceQueueBuffers call will change the source type to ALSourceType.Streaming.
+        /// sequence, and the number of processed buffers can be detected using <see cref="AL.GetSource"/> with parameter
+        /// <see cref="ALGetSourcei.BuffersProcessed"/>. When first created, a source will be of type <see cref="ALSourceType.Undetermined"/>. A successful
+        /// <see cref="AL.SourceQueueBuffers"/> call will change the source type to <see cref="ALSourceType.Streaming"/>.
         /// </summary>
         /// <param name="sid">The name of the source to queue buffers onto.</param>
         /// <param name="numEntries">The number of buffers to be queued.</param>
@@ -1320,9 +1135,9 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function queues a set of buffers on a source. All buffers attached to a source will be played in
-        /// sequence, and the number of processed buffers can be detected using AL.GetSource with parameter
-        /// ALGetSourcei.BuffersProcessed. When first created, a source will be of type ALSourceType.Undetermined. A successful
-        /// AL.SourceQueueBuffers call will change the source type to ALSourceType.Streaming.
+        /// sequence, and the number of processed buffers can be detected using <see cref="AL.GetSource"/> with parameter
+        /// <see cref="ALGetSourcei.BuffersProcessed"/>. When first created, a source will be of type <see cref="ALSourceType.Undetermined"/>. A successful
+        /// <see cref="AL.SourceQueueBuffers"/> call will change the source type to <see cref="ALSourceType.Streaming"/>.
         /// </summary>
         /// <param name="sid">The name of the source to queue buffers onto.</param>
         /// <param name="numEntries">The number of buffers to be queued.</param>
@@ -1340,9 +1155,9 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function queues a set of buffers on a source. All buffers attached to a source will be played in
-        /// sequence, and the number of processed buffers can be detected using AL.GetSource with parameter
-        /// ALGetSourcei.BuffersProcessed. When first created, a source will be of type ALSourceType.Undetermined. A successful
-        /// AL.SourceQueueBuffers call will change the source type to ALSourceType.Streaming.
+        /// sequence, and the number of processed buffers can be detected using <see cref="AL.GetSource"/> with parameter
+        /// <see cref="ALGetSourcei.BuffersProcessed"/>. When first created, a source will be of type <see cref="ALSourceType.Undetermined"/>. A successful
+        /// <see cref="AL.SourceQueueBuffers"/> call will change the source type to <see cref="ALSourceType.Streaming"/>.
         /// </summary>
         /// <param name="sid">The name of the source to queue buffers onto.</param>
         /// <param name="numEntries">The number of buffers to be queued.</param>
@@ -1360,9 +1175,9 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function queues a set of buffers on a source. All buffers attached to a source will be played in
-        /// sequence, and the number of processed buffers can be detected using AL.GetSource with parameter
-        /// ALGetSourcei.BuffersProcessed. When first created, a source will be of type ALSourceType.Undetermined. A successful
-        /// AL.SourceQueueBuffers call will change the source type to ALSourceType.Streaming.
+        /// sequence, and the number of processed buffers can be detected using <see cref="AL.GetSource"/> with parameter
+        /// <see cref="ALGetSourcei.BuffersProcessed"/>. When first created, a source will be of type <see cref="ALSourceType.Undetermined"/>. A successful
+        /// <see cref="AL.SourceQueueBuffers"/> call will change the source type to <see cref="ALSourceType.Streaming"/>.
         /// </summary>
         /// <param name="source">The name of the source to queue buffers onto.</param>
         /// <param name="buffer">The name of the buffer to be queued.</param>
@@ -1376,77 +1191,77 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function unqueues a set of buffers attached to a source. The number of processed buffers can be detected
-        /// using AL.GetSource with parameter ALGetSourcei.BuffersProcessed, which is the maximum number of buffers that can be
+        /// using <see cref="AL.GetSource"/> with parameter <see cref="ALGetSourcei.BuffersProcessed"/>, which is the maximum number of buffers that can be
         /// unqueued using this call. The unqueue operation will only take place if all n buffers can be removed from the
         /// queue.
         /// </summary>
         /// <param name="sid">The name of the source to unqueue buffers from.</param>
         /// <param name="numEntries">The number of buffers to be unqueued.</param>
         /// <param name="bids">A pointer to an array of buffer names that were removed.</param>
-        [DllImport(Lib, EntryPoint = "alSourceUnqueueBuffers")]
+        [DllImport(Library, EntryPoint = "alSourceUnqueueBuffers")]
         [SuppressUnmanagedCodeSecurity]
-        public static extern unsafe void SourceUnqueueBuffers(uint sid, int numEntries, [In] uint* bids);
-        // AL_API void AL_APIENTRY alSourceUnqueueBuffers( ALuint sid, ALsizei numEntries, ALuint *bids );
+        public static extern unsafe void SourceUnqueueBuffers(uint sid, int numEntries, [Out] uint* bids);
 
         /// <summary>
         /// This function unqueues a set of buffers attached to a source. The number of processed buffers can be detected
-        /// using AL.GetSource with parameter ALGetSourcei.BuffersProcessed, which is the maximum number of buffers that can be
+        /// using <see cref="AL.GetSource"/> with parameter <see cref="ALGetSourcei.BuffersProcessed"/>, which is the maximum number of buffers that can be
         /// unqueued using this call. The unqueue operation will only take place if all n buffers can be removed from the
         /// queue.
         /// </summary>
         /// <param name="sid">The name of the source to unqueue buffers from.</param>
         /// <param name="numEntries">The number of buffers to be unqueued.</param>
         /// <param name="bids">A pointer to an array of buffer names that were removed.</param>
-        [DllImport(Lib, EntryPoint = "alSourceUnqueueBuffers")]
+        [DllImport(Library, EntryPoint = "alSourceUnqueueBuffers")]
         [SuppressUnmanagedCodeSecurity]
         public static extern void SourceUnqueueBuffers(uint sid, int numEntries, [Out] uint[] bids);
 
         /// <summary>
         /// This function unqueues a set of buffers attached to a source. The number of processed buffers can be detected
-        /// using AL.GetSource with parameter ALGetSourcei.BuffersProcessed, which is the maximum number of buffers that can be
+        /// using <see cref="AL.GetSource"/> with parameter <see cref="ALGetSourcei.BuffersProcessed"/>, which is the maximum number of buffers that can be
         /// unqueued using this call. The unqueue operation will only take place if all n buffers can be removed from the
         /// queue.
         /// </summary>
         /// <param name="sid">The name of the source to unqueue buffers from.</param>
         /// <param name="numEntries">The number of buffers to be unqueued.</param>
         /// <param name="bids">A pointer to an array of buffer names that were removed.</param>
-        [DllImport(Lib, EntryPoint = "alSourceUnqueueBuffers")]
+        [DllImport(Library, EntryPoint = "alSourceUnqueueBuffers")]
         [SuppressUnmanagedCodeSecurity]
         public static extern void SourceUnqueueBuffers(int sid, int numEntries, [Out] int[] bids);
 
         /// <summary>
         /// This function unqueues a set of buffers attached to a source. The number of processed buffers can be detected
-        /// using AL.GetSource with parameter ALGetSourcei.BuffersProcessed, which is the maximum number of buffers that can be
+        /// using <see cref="AL.GetSource"/> with parameter <see cref="ALGetSourcei.BuffersProcessed"/>, which is the maximum number of buffers that can be
         /// unqueued using this call. The unqueue operation will only take place if all n buffers can be removed from the
         /// queue.
         /// </summary>
         /// <param name="sid">The name of the source to unqueue buffers from.</param>
         /// <param name="numEntries">The number of buffers to be unqueued.</param>
         /// <param name="bids">A pointer to an array of buffer names that were removed.</param>
-        [DllImport(Lib, EntryPoint = "alSourceUnqueueBuffers")]
+        [DllImport(Library, EntryPoint = "alSourceUnqueueBuffers")]
         [SuppressUnmanagedCodeSecurity]
-        public static extern void SourceUnqueueBuffers(uint sid, int numEntries, ref uint bids);
+        public static extern void SourceUnqueueBuffers(uint sid, int numEntries, out uint bids);
 
         /// <summary>
         /// This function unqueues a set of buffers attached to a source. The number of processed buffers can be detected
-        /// using AL.GetSource with parameter ALGetSourcei.BuffersProcessed, which is the maximum number of buffers that can be
+        /// using <see cref="AL.GetSource"/> with parameter <see cref="ALGetSourcei.BuffersProcessed"/>, which is the maximum number of buffers that can be
         /// unqueued using this call. The unqueue operation will only take place if all n buffers can be removed from the
         /// queue.
         /// </summary>
         /// <param name="sid">The name of the source to unqueue buffers from.</param>
         /// <param name="numEntries">The number of buffers to be unqueued.</param>
         /// <param name="bids">A pointer to an array of buffer names that were removed.</param>
-        [DllImport(Lib, EntryPoint = "alSourceUnqueueBuffers")]
+        [DllImport(Library, EntryPoint = "alSourceUnqueueBuffers")]
         [SuppressUnmanagedCodeSecurity]
-        public static extern void SourceUnqueueBuffers(int sid, int numEntries, ref int bids);
+        public static extern void SourceUnqueueBuffers(int sid, int numEntries, out int bids);
 
         /// <summary>
         /// This function unqueues a set of buffers attached to a source. The number of processed buffers can be detected
-        /// using AL.GetSource with parameter ALGetSourcei.BuffersProcessed, which is the maximum number of buffers that can be
+        /// using <see cref="AL.GetSource"/> with parameter <see cref="ALGetSourcei.BuffersProcessed"/>, which is the maximum number of buffers that can be
         /// unqueued using this call. The unqueue operation will only take place if all n buffers can be removed from the
         /// queue.
         /// </summary>
         /// <param name="sid">The name of the source to unqueue buffers from.</param>
+        /// <returns>The buffer that was unqueued.</returns>
         public static int SourceUnqueueBuffer(int sid)
         {
             uint buf;
@@ -1460,12 +1275,13 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function unqueues a set of buffers attached to a source. The number of processed buffers can be detected
-        /// using AL.GetSource with parameter ALGetSourcei.BuffersProcessed, which is the maximum number of buffers that can be
+        /// using <see cref="AL.GetSource"/> with parameter <see cref="ALGetSourcei.BuffersProcessed"/>, which is the maximum number of buffers that can be
         /// unqueued using this call. The unqueue operation will only take place if all n buffers can be removed from the
         /// queue.
         /// </summary>
         /// <param name="sid">The name of the source to unqueue buffers from.</param>
         /// <param name="numEntries">The number of buffers to be unqueued.</param>
+        /// <returns>The buffers that were unqueued.</returns>
         public static int[] SourceUnqueueBuffers(int sid, int numEntries)
         {
             if (numEntries <= 0)
@@ -1493,32 +1309,31 @@ namespace OpenTK.OpenAL.Native
          */
 
         /// <summary>
-        /// This function generates one or more buffers, which contain audio buffer (see AL.BufferData). References to
-        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as AL.DeleteBuffers,
-        /// AL.Source with parameter ALSourcei, AL.SourceQueueBuffers, and AL.SourceUnqueueBuffers).
+        /// This function generates one or more buffers, which contain audio buffer (see <see cref="AL.BufferData"/>). References to
+        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as <see cref="AL.DeleteBuffers"/>,
+        /// <see cref="AL.Source"/> with parameter ALSourcei, <see cref="AL.SourceQueueBuffers"/>, and <see cref="AL.SourceUnqueueBuffers"/>).
         /// </summary>
         /// <param name="n">The number of buffers to be generated.</param>
         /// <param name="buffers">Pointer to an array of uint values which will store the names of the new buffers.</param>
-        [DllImport(Lib, EntryPoint = "alGenBuffers", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGenBuffers", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void GenBuffers(int n, [Out] uint* buffers);
-        // AL_API void AL_APIENTRY alGenBuffers( ALsizei n, ALuint* Buffers );
 
         /// <summary>
-        /// This function generates one or more buffers, which contain audio buffer (see AL.BufferData). References to
-        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as AL.DeleteBuffers,
-        /// AL.Source with parameter ALSourcei, AL.SourceQueueBuffers, and AL.SourceUnqueueBuffers).
+        /// This function generates one or more buffers, which contain audio buffer (see <see cref="AL.BufferData"/>). References to
+        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as <see cref="AL.DeleteBuffers"/>,
+        /// <see cref="AL.Source"/> with parameter ALSourcei, <see cref="AL.SourceQueueBuffers"/>, and <see cref="AL.SourceUnqueueBuffers"/>).
         /// </summary>
         /// <param name="n">The number of buffers to be generated.</param>
         /// <param name="buffers">Pointer to an array of uint values which will store the names of the new buffers.</param>
-        [DllImport(Lib, EntryPoint = "alGenBuffers", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alGenBuffers", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void GenBuffers(int n, [Out] int* buffers);
 
         /// <summary>
-        /// This function generates one or more buffers, which contain audio buffer (see AL.BufferData). References to
-        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as AL.DeleteBuffers,
-        /// AL.Source with parameter ALSourcei, AL.SourceQueueBuffers, and AL.SourceUnqueueBuffers).
+        /// This function generates one or more buffers, which contain audio buffer (see <see cref="AL.BufferData"/>). References to
+        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as <see cref="AL.DeleteBuffers"/>,
+        /// <see cref="AL.Source"/> with parameter ALSourcei, <see cref="AL.SourceQueueBuffers"/>, and <see cref="AL.SourceUnqueueBuffers"/>).
         /// </summary>
         /// <param name="n">The number of buffers to be generated.</param>
         /// <param name="buffers">Pointer to an array of uint values which will store the names of the new buffers.</param>
@@ -1534,9 +1349,9 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function generates one or more buffers, which contain audio buffer (see AL.BufferData). References to
-        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as AL.DeleteBuffers,
-        /// AL.Source with parameter ALSourcei, AL.SourceQueueBuffers, and AL.SourceUnqueueBuffers).
+        /// This function generates one or more buffers, which contain audio buffer (see <see cref="AL.BufferData"/>). References to
+        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as <see cref="AL.DeleteBuffers"/>,
+        /// <see cref="AL.Source"/> with parameter ALSourcei, <see cref="AL.SourceQueueBuffers"/>, and <see cref="AL.SourceUnqueueBuffers"/>).
         /// </summary>
         /// <param name="n">The number of buffers to be generated.</param>
         /// <param name="buffers">Pointer to an array of uint values which will store the names of the new buffers.</param>
@@ -1552,9 +1367,9 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function generates one or more buffers, which contain audio data (see AL.BufferData). References to
-        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as AL.DeleteBuffers,
-        /// AL.Source with parameter ALSourcei, AL.SourceQueueBuffers, and AL.SourceUnqueueBuffers).
+        /// This function generates one or more buffers, which contain audio data (see <see cref="AL.BufferData"/>). References to
+        /// buffers are uint values, which are used wherever a buffer reference is needed (in calls such as <see cref="AL.DeleteBuffers"/>,
+        /// <see cref="AL.Source"/> with parameter ALSourcei, <see cref="AL.SourceQueueBuffers"/>, and <see cref="AL.SourceUnqueueBuffers"/>).
         /// </summary>
         /// <param name="n">The number of buffers to be generated.</param>
         /// <returns>Pointer to an array of uint values which will store the names of the new buffers.</returns>
@@ -1566,9 +1381,9 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function generates one buffer only, which contain audio data (see AL.BufferData). References to buffers
-        /// are uint values, which are used wherever a buffer reference is needed (in calls such as AL.DeleteBuffers, AL.Source
-        /// with parameter ALSourcei, AL.SourceQueueBuffers, and AL.SourceUnqueueBuffers).
+        /// This function generates one buffer only, which contain audio data (see <see cref="AL.BufferData"/>). References to buffers
+        /// are uint values, which are used wherever a buffer reference is needed (in calls such as <see cref="AL.DeleteBuffers"/>, <see cref="AL.Source"/>
+        /// with parameter ALSourcei, <see cref="AL.SourceQueueBuffers"/>, and <see cref="AL.SourceUnqueueBuffers"/>).
         /// </summary>
         /// <returns>Pointer to an uint value which will store the name of the new buffer.</returns>
         public static int GenBuffer()
@@ -1578,9 +1393,9 @@ namespace OpenTK.OpenAL.Native
         }
 
         /// <summary>
-        /// This function generates one buffer only, which contain audio data (see AL.BufferData). References to buffers
-        /// are uint values, which are used wherever a buffer reference is needed (in calls such as AL.DeleteBuffers, AL.Source
-        /// with parameter ALSourcei, AL.SourceQueueBuffers, and AL.SourceUnqueueBuffers).
+        /// This function generates one buffer only, which contain audio data (see <see cref="AL.BufferData"/>). References to buffers
+        /// are uint values, which are used wherever a buffer reference is needed (in calls such as <see cref="AL.DeleteBuffers"/>, <see cref="AL.Source"/>
+        /// with parameter ALSourcei, <see cref="AL.SourceQueueBuffers"/>, and <see cref="AL.SourceUnqueueBuffers"/>).
         /// </summary>
         /// <param name="buffer">Pointer to an uint value which will store the names of the new buffer.</param>
         public static void GenBuffer(out uint buffer)
@@ -1590,30 +1405,29 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function deletes one or more buffers, freeing the resources used by the buffer. Buffers which are
-        /// attached to a source can not be deleted. See AL.Source (ALSourcei) and AL.SourceUnqueueBuffers for information on
+        /// attached to a source can not be deleted. See <see cref="AL.Source"/> (ALSourcei) and <see cref="AL.SourceUnqueueBuffers"/> for information on
         /// how to detach a buffer from a source.
         /// </summary>
         /// <param name="n">The number of buffers to be deleted.</param>
         /// <param name="buffers">Pointer to an array of buffer names identifying the buffers to be deleted.</param>
-        [DllImport(Lib, EntryPoint = "alDeleteBuffers", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alDeleteBuffers", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void DeleteBuffers(int n, [In] uint* buffers);
-        // AL_API void AL_APIENTRY alDeleteBuffers( ALsizei n, const ALuint* Buffers );
 
         /// <summary>
         /// This function deletes one or more buffers, freeing the resources used by the buffer. Buffers which are
-        /// attached to a source can not be deleted. See AL.Source (ALSourcei) and AL.SourceUnqueueBuffers for information on
+        /// attached to a source can not be deleted. See <see cref="AL.Source"/> (ALSourcei) and <see cref="AL.SourceUnqueueBuffers"/> for information on
         /// how to detach a buffer from a source.
         /// </summary>
         /// <param name="n">The number of buffers to be deleted.</param>
         /// <param name="buffers">Pointer to an array of buffer names identifying the buffers to be deleted.</param>
-        [DllImport(Lib, EntryPoint = "alDeleteBuffers", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alDeleteBuffers", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern unsafe void DeleteBuffers(int n, [In] int* buffers);
 
         /// <summary>
         /// This function deletes one or more buffers, freeing the resources used by the buffer. Buffers which are
-        /// attached to a source can not be deleted. See AL.Source (ALSourcei) and AL.SourceUnqueueBuffers for information on
+        /// attached to a source can not be deleted. See <see cref="AL.Source"/> (ALSourcei) and <see cref="AL.SourceUnqueueBuffers"/> for information on
         /// how to detach a buffer from a source.
         /// </summary>
         /// <param name="n">The number of buffers to be deleted.</param>
@@ -1631,7 +1445,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function deletes one or more buffers, freeing the resources used by the buffer. Buffers which are
-        /// attached to a source can not be deleted. See AL.Source (ALSourcei) and AL.SourceUnqueueBuffers for information on
+        /// attached to a source can not be deleted. See <see cref="AL.Source"/> (ALSourcei) and <see cref="AL.SourceUnqueueBuffers"/> for information on
         /// how to detach a buffer from a source.
         /// </summary>
         /// <param name="n">The number of buffers to be deleted.</param>
@@ -1649,7 +1463,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function deletes one buffer only, freeing the resources used by the buffer. Buffers which are attached to
-        /// a source can not be deleted. See AL.Source (ALSourcei) and AL.SourceUnqueueBuffers for information on how to detach
+        /// a source can not be deleted. See <see cref="AL.Source"/> (ALSourcei) and <see cref="AL.SourceUnqueueBuffers"/> for information on how to detach
         /// a buffer from a source.
         /// </summary>
         /// <param name="buffers">Pointer to a buffer name identifying the buffer to be deleted.</param>
@@ -1670,7 +1484,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function deletes one or more buffers, freeing the resources used by the buffer. Buffers which are
-        /// attached to a source can not be deleted. See AL.Source (ALSourcei) and AL.SourceUnqueueBuffers for information on
+        /// attached to a source can not be deleted. See <see cref="AL.Source"/> (ALSourcei) and <see cref="AL.SourceUnqueueBuffers"/> for information on
         /// how to detach a buffer from a source.
         /// </summary>
         /// <param name="buffers">Pointer to an array of buffer names identifying the buffers to be deleted.</param>
@@ -1691,7 +1505,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function deletes one buffer only, freeing the resources used by the buffer. Buffers which are attached to
-        /// a source can not be deleted. See AL.Source (ALSourcei) and AL.SourceUnqueueBuffers for information on how to detach
+        /// a source can not be deleted. See <see cref="AL.Source"/> (ALSourcei) and <see cref="AL.SourceUnqueueBuffers"/> for information on how to detach
         /// a buffer from a source.
         /// </summary>
         /// <param name="buffer">Pointer to a buffer name identifying the buffer to be deleted.</param>
@@ -1702,7 +1516,7 @@ namespace OpenTK.OpenAL.Native
 
         /// <summary>
         /// This function deletes one buffer only, freeing the resources used by the buffer. Buffers which are attached to
-        /// a source can not be deleted. See AL.Source (ALSourcei) and AL.SourceUnqueueBuffers for information on how to detach
+        /// a source can not be deleted. See <see cref="AL.Source"/> (ALSourcei) and <see cref="AL.SourceUnqueueBuffers"/> for information on how to detach
         /// a buffer from a source.
         /// </summary>
         /// <param name="buffer">Pointer to a buffer name identifying the buffer to be deleted.</param>
@@ -1716,10 +1530,9 @@ namespace OpenTK.OpenAL.Native
         /// </summary>
         /// <param name="bid">A buffer Handle previously allocated with <see cref="GenBuffers(int)" />.</param>
         /// <returns>Success.</returns>
-        [DllImport(Lib, EntryPoint = "alIsBuffer", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alIsBuffer", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern bool IsBuffer(uint bid);
-        // AL_API ALboolean AL_APIENTRY alIsBuffer( ALuint bid );
 
         /// <summary>
         /// This function tests if a buffer name is valid, returning True if valid, False if not.
@@ -1738,16 +1551,15 @@ namespace OpenTK.OpenAL.Native
         /// </summary>
         /// <param name="bid">buffer Handle/Name to be filled with buffer.</param>
         /// <param name="format">
-        /// Format type from among the following: ALFormat.Mono8, ALFormat.Mono16, ALFormat.Stereo8,
-        /// ALFormat.Stereo16.
+        /// Format type from among the following: <see cref="ALFormat.Mono8"/>, <see cref="ALFormat.Mono16"/>, <see cref="ALFormat.Stereo8"/>,
+        /// <see cref="ALFormat.Stereo16"/>.
         /// </param>
         /// <param name="buffer">Pointer to a pinned audio buffer.</param>
         /// <param name="size">The size of the audio buffer in bytes.</param>
         /// <param name="freq">The frequency of the audio buffer.</param>
-        [DllImport(Lib, EntryPoint = "alBufferData", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alBufferData", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void BufferData(uint bid, ALFormat format, IntPtr buffer, int size, int freq);
-        // AL_API void AL_APIENTRY alBufferData( ALuint bid, ALenum format, const ALvoid* buffer, ALsizei size, ALsizei freq );
 
         /// <summary>
         /// This function fills a buffer with audio buffer. All the pre-defined formats are PCM buffer, but this function
@@ -1755,8 +1567,8 @@ namespace OpenTK.OpenAL.Native
         /// </summary>
         /// <param name="bid">buffer Handle/Name to be filled with buffer.</param>
         /// <param name="format">
-        /// Format type from among the following: ALFormat.Mono8, ALFormat.Mono16, ALFormat.Stereo8,
-        /// ALFormat.Stereo16.
+        /// Format type from among the following: <see cref="ALFormat.Mono8"/>, <see cref="ALFormat.Mono16"/>, <see cref="ALFormat.Stereo8"/>,
+        /// <see cref="ALFormat.Stereo16"/>.
         /// </param>
         /// <param name="buffer">Pointer to a pinned audio buffer.</param>
         /// <param name="size">The size of the audio buffer in bytes.</param>
@@ -1772,9 +1584,10 @@ namespace OpenTK.OpenAL.Native
         /// </summary>
         /// <param name="bid">buffer Handle/Name to be filled with buffer.</param>
         /// <param name="format">
-        /// Format type from among the following: ALFormat.Mono8, ALFormat.Mono16, ALFormat.Stereo8,
-        /// ALFormat.Stereo16.
+        /// Format type from among the following: <see cref="ALFormat.Mono8"/>, <see cref="ALFormat.Mono16"/>, <see cref="ALFormat.Stereo8"/>,
+        /// <see cref="ALFormat.Stereo16"/>.
         /// </param>
+        /// <typeparam name="TBuffer">The type of the buffer.</typeparam>
         /// <param name="buffer">The audio buffer.</param>
         /// <param name="size">The size of the audio buffer in bytes.</param>
         /// <param name="freq">The frequency of the audio buffer.</param>
@@ -1797,136 +1610,103 @@ namespace OpenTK.OpenAL.Native
             }
         }
 
-        /*
-        Remarks (from Manual)
-         * There are no relevant buffer properties defined in OpenAL 1.1 which can be affected by this call,
-         * but this function may be used by OpenAL extensions.
-
-        // AL_API void AL_APIENTRY alBufferf( ALuint bid, ALenum param, ALfloat value );
-        // AL_API void AL_APIENTRY alBufferfv( ALuint bid, ALenum param, const ALfloat* values );
-        // AL_API void AL_APIENTRY alBufferi( ALuint bid, ALenum param, ALint value );
-        // AL_API void AL_APIENTRY alBuffer3i( ALuint bid, ALenum param, ALint value1, ALint value2, ALint value3 );
-        // AL_API void AL_APIENTRY alBufferiv( ALuint bid, ALenum param, const ALint* values );
-        // AL_API void AL_APIENTRY alBuffer3f( ALuint bid, ALenum param, ALfloat value1, ALfloat value2, ALfloat value3 );
-        */
-
-        /*
-        [DllImport( Al.Lib, EntryPoint = "alBuffer3f", ExactSpelling = true, CallingConvention = Al.Style ), SuppressUnmanagedCodeSecurity( )]
-        public static extern void Buffer3f( uint bid, ALenum param, ALfloat value1, ALfloat value2, ALfloat value3 );
-
-        public static void Bufferv3( uint bid, Alenum param, ref Vector3 values )
-        {
-            Buffer3f( bid, param, values.X, values.Y, values.Z );
-        }*/
-
         /// <summary>
         /// This function retrieves an integer property of a buffer.
         /// </summary>
-        /// <param name="bid">Buffer name whose attribute is being retrieved</param>
+        /// <param name="bid">Buffer name whose attribute is being retrieved.</param>
         /// <param name="param">
-        /// The name of the attribute to be retrieved: ALGetBufferi.Frequency, Bits, Channels, Size, and the
+        /// The name of the attribute to be retrieved: <see cref="ALGetBufferi.Frequency"/>, Bits, Channels, Size, and the
         /// currently hidden AL_DATA (dangerous).
         /// </param>
-        /// <param name="value">A pointer to an int to hold the retrieved buffer</param>
-        [DllImport(Lib, EntryPoint = "alGetBufferi", ExactSpelling = true, CallingConvention = Style)]
+        /// <param name="value">A pointer to an int to hold the retrieved buffer.</param>
+        [DllImport(Library, EntryPoint = "alGetBufferi", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void GetBuffer(uint bid, ALGetBufferi param, [Out] out int value);
-        // AL_API void AL_APIENTRY alGetBufferi( ALuint bid, ALenum param, ALint* value );
 
         /// <summary>
         /// This function retrieves an integer property of a buffer.
         /// </summary>
-        /// <param name="bid">Buffer name whose attribute is being retrieved</param>
+        /// <param name="bid">Buffer name whose attribute is being retrieved.</param>
         /// <param name="param">
-        /// The name of the attribute to be retrieved: ALGetBufferi.Frequency, Bits, Channels, Size, and the
+        /// The name of the attribute to be retrieved: <see cref="ALGetBufferi.Frequency"/>, Bits, Channels, Size, and the
         /// currently hidden AL_DATA (dangerous).
         /// </param>
-        /// <param name="value">A pointer to an int to hold the retrieved buffer</param>
+        /// <param name="value">A pointer to an int to hold the retrieved buffer.</param>
         public static void GetBuffer(int bid, ALGetBufferi param, out int value)
         {
             GetBuffer((uint)bid, param, out value);
         }
 
-        // AL_API void AL_APIENTRY alGetBufferf( ALuint bid, ALenum param, ALfloat* value );
-        // AL_API void AL_APIENTRY alGetBuffer3f( ALuint bid, ALenum param, ALfloat* value1, ALfloat* value2, ALfloat* value3);
-        // AL_API void AL_APIENTRY alGetBufferfv( ALuint bid, ALenum param, ALfloat* values );
-        // AL_API void AL_APIENTRY alGetBuffer3i( ALuint bid, ALenum param, ALint* value1, ALint* value2, ALint* value3);
-        // AL_API void AL_APIENTRY alGetBufferiv( ALuint bid, ALenum param, ALint* values );
-
         /// <summary>
-        /// AL.DopplerFactor is a simple scaling of source and listener velocities to exaggerate or deemphasize the
+        /// <see cref="AL.DopplerFactor"/> is a simple scaling of source and listener velocities to exaggerate or deemphasize the
         /// Doppler (pitch) shift resulting from the calculation.
         /// </summary>
         /// <param name="value">
         /// A negative value will result in an error, the command is then ignored. The default value is 1f. The
-        /// current setting can be queried using AL.Get with parameter ALGetFloat.SpeedOfSound.
+        /// current setting can be queried using <see cref="AL.Get"/> with parameter <see cref="ALGetFloat.SpeedOfSound"/>.
         /// </param>
-        [DllImport(Lib, EntryPoint = "alDopplerFactor", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alDopplerFactor", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void DopplerFactor(float value);
-        // AL_API void AL_APIENTRY alDopplerFactor( ALfloat value );
 
         /// <summary>
         /// This function is deprecated and should not be used.
         /// </summary>
         /// <param name="value">The default is 1.0f.</param>
-        [DllImport(Lib, EntryPoint = "alDopplerVelocity", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alDopplerVelocity", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void DopplerVelocity(float value);
-        // AL_API void AL_APIENTRY alDopplerVelocity( ALfloat value );
 
         /// <summary>
-        /// AL.SpeedOfSound allows the application to change the reference (propagation) speed used in the Doppler
+        /// <see cref="AL.SpeedOfSound"/> allows the application to change the reference (propagation) speed used in the Doppler
         /// calculation. The source and listener velocities should be expressed in the same units as the speed of sound.
         /// </summary>
         /// <param name="value">
         /// A negative or zero value will result in an error, and the command is ignored. Default: 343.3f
         /// (appropriate for velocity units of meters and air as the propagation medium). The current setting can be queried
-        /// using AL.Get with parameter ALGetFloat.SpeedOfSound.
+        /// using <see cref="AL.Get"/> with parameter <see cref="ALGetFloat.SpeedOfSound"/>.
         /// </param>
-        [DllImport(Lib, EntryPoint = "alSpeedOfSound", ExactSpelling = true, CallingConvention = Style)]
+        [DllImport(Library, EntryPoint = "alSpeedOfSound", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void SpeedOfSound(float value);
-        // AL_API void AL_APIENTRY alSpeedOfSound( ALfloat value );
 
         /// <summary>
-        /// This function selects the OpenAL distance model – ALDistanceModel.InverseDistance,
-        /// ALDistanceModel.InverseDistanceClamped, ALDistanceModel.LinearDistance, ALDistanceModel.LinearDistanceClamped,
-        /// ALDistanceModel.ExponentDistance, ALDistanceModel.ExponentDistanceClamped, or ALDistanceModel.None. The default
-        /// distance model in OpenAL is ALDistanceModel.InverseDistanceClamped.
+        /// This function selects the OpenAL distance model – <see cref="ALDistanceModel.InverseDistance"/>,
+        /// <see cref="ALDistanceModel.InverseDistanceClamped"/>, <see cref="ALDistanceModel.LinearDistance"/>, <see cref="ALDistanceModel.LinearDistanceClamped"/>,
+        /// <see cref="ALDistanceModel.ExponentDistance"/>, <see cref="ALDistanceModel.ExponentDistanceClamped"/>, or <see cref="ALDistanceModel.None"/>. The default
+        /// distance model in OpenAL is <see cref="ALDistanceModel.InverseDistanceClamped"/>.
         /// </summary>
         /// <remarks>
         /// The ALDistanceModel .InverseDistance model works according to the following formula:
-        /// gain = ALSourcef.ReferenceDistance / (ALSourcef.ReferenceDistance + ALSourcef.RolloffFactor * (distance –
-        /// ALSourcef.ReferenceDistance));
+        /// gain = <see cref="ALSourcef.ReferenceDistance"/> / (<see cref="ALSourcef.ReferenceDistance"/> + <see cref="ALSourcef.RolloffFactor"/> * (distance –
+        /// <see cref="ALSourcef.ReferenceDistance"/>));
         /// The ALDistanceModel .InverseDistanceClamped model works according to the following formula:
-        /// distance = max(distance,ALSourcef.ReferenceDistance);
-        /// distance = min(distance,ALSourcef.MaxDistance);
-        /// gain = ALSourcef.ReferenceDistance / (ALSourcef.ReferenceDistance + ALSourcef.RolloffFactor * (distance –
-        /// ALSourcef.ReferenceDistance));
-        /// The ALDistanceModel.LinearDistance model works according to the following formula:
-        /// distance = min(distance, ALSourcef.MaxDistance) // avoid negative gain
-        /// gain = (1 – ALSourcef.RolloffFactor * (distance – ALSourcef.ReferenceDistance) / (ALSourcef.MaxDistance –
-        /// ALSourcef.ReferenceDistance))
-        /// The ALDistanceModel.LinearDistanceClamped model works according to the following formula:
-        /// distance = max(distance, ALSourcef.ReferenceDistance)
-        /// distance = min(distance, ALSourcef.MaxDistance)
-        /// gain = (1 – ALSourcef.RolloffFactor * (distance – ALSourcef.ReferenceDistance) / (ALSourcef.MaxDistance –
-        /// ALSourcef.ReferenceDistance))
-        /// The ALDistanceModel.ExponentDistance model works according to the following formula:
-        /// gain = (distance / ALSourcef.ReferenceDistance) ^ (- ALSourcef.RolloffFactor)
-        /// The ALDistanceModel.ExponentDistanceClamped model works according to the following formula:
-        /// distance = max(distance, ALSourcef.ReferenceDistance)
-        /// distance = min(distance, ALSourcef.MaxDistance)
-        /// gain = (distance / ALSourcef.ReferenceDistance) ^ (- ALSourcef.RolloffFactor)
-        /// The ALDistanceModel.None model works according to the following formula:
-        /// gain = 1f;
+        /// distance = max(distance,<see cref="ALSourcef.ReferenceDistance"/>);
+        /// distance = min(distance,<see cref="ALSourcef.MaxDistance"/>);
+        /// gain = <see cref="ALSourcef.ReferenceDistance"/> / (<see cref="ALSourcef.ReferenceDistance"/> + <see cref="ALSourcef.RolloffFactor"/> * (distance –
+        /// <see cref="ALSourcef.ReferenceDistance"/>));
+        /// The <see cref="ALDistanceModel.LinearDistance"/> model works according to the following formula:
+        /// distance = min(distance, <see cref="ALSourcef.MaxDistance"/>) // avoid negative gain
+        /// gain = (1 – <see cref="ALSourcef.RolloffFactor"/> * (distance – <see cref="ALSourcef.ReferenceDistance"/>) / (<see cref="ALSourcef.MaxDistance"/> –
+        /// <see cref="ALSourcef.ReferenceDistance"/>))
+        /// The <see cref="ALDistanceModel.LinearDistanceClamped"/> model works according to the following formula:
+        /// distance = max(distance, <see cref="ALSourcef.ReferenceDistance"/>)
+        /// distance = min(distance, <see cref="ALSourcef.MaxDistance"/>)
+        /// gain = (1 – <see cref="ALSourcef.RolloffFactor"/> * (distance – <see cref="ALSourcef.ReferenceDistance"/>) / (<see cref="ALSourcef.MaxDistance"/> –
+        /// <see cref="ALSourcef.ReferenceDistance"/>))
+        /// The <see cref="ALDistanceModel.ExponentDistance"/> model works according to the following formula:
+        /// gain = (distance / <see cref="ALSourcef.ReferenceDistance"/>) ^ (- <see cref="ALSourcef.RolloffFactor"/>)
+        /// The <see cref="ALDistanceModel.ExponentDistanceClamped"/> model works according to the following formula:
+        /// distance = max(distance, <see cref="ALSourcef.ReferenceDistance"/>)
+        /// distance = min(distance, <see cref="ALSourcef.MaxDistance"/>)
+        /// gain = (distance / <see cref="ALSourcef.ReferenceDistance"/>) ^ (- <see cref="ALSourcef.RolloffFactor"/>)
+        /// The <see cref="ALDistanceModel.None"/> model works according to the following formula:
+        /// gain = 1f.
         /// </remarks>
-        /// <param name="distancemodel"></param>
-        [DllImport(Lib, EntryPoint = "alDistanceModel", ExactSpelling = true, CallingConvention = Style)]
+        /// <param name="distancemodel">The distance model.</param>
+        [DllImport(Library, EntryPoint = "alDistanceModel", ExactSpelling = true, CallingConvention = Style)]
         [SuppressUnmanagedCodeSecurity]
         public static extern void DistanceModel(ALDistanceModel distancemodel);
-        // AL_API void AL_APIENTRY alDistanceModel( ALenum distanceModel );
 
         /// <summary>
         /// (Helper) Returns Source state information.
