+++ src/OpenTK.OpenAL/Native/AL/Enums/ALCapability.cs
@@ -1,7 +1,7 @@
 ﻿namespace OpenTK.OpenAL.Native
 {
     /// <summary>
-    /// A list of valid Enable/Disable/IsEnabled parameters
+    /// A list of valid Enable/Disable/IsEnabled parameters.
     /// </summary>
     public enum ALCapability
     {
