+++ src/OpenTK.OpenAL/Native/AL/Enums/ALGetFloat.cs
@@ -1,12 +1,12 @@
 ﻿namespace OpenTK.OpenAL.Native
 {
     /// <summary>
-    /// A list of valid 32-bit Float AL.Get() parameters
+    /// A list of valid 32-bit Float AL.Get() parameters.
     /// </summary>
     public enum ALGetFloat
     {
         /// <summary>
-        /// Doppler scale. Default 1.0f
+        /// Doppler scale. Default 1.0f.
         /// </summary>
         DopplerFactor = 0xC000,
 
@@ -16,7 +16,7 @@
         DopplerVelocity = 0xC001,
 
         /// <summary>
-        /// Speed of Sound in units per second. Default: 343.3f
+        /// Speed of Sound in units per second. Default: 343.3f.
         /// </summary>
         SpeedOfSound = 0xC003
     }
