+++ src/OpenTK.OpenAL/Native/AL/Enums/ALGetString.cs
@@ -1,7 +1,7 @@
 ﻿namespace OpenTK.OpenAL.Native
 {
     /// <summary>
-    /// A list of valid string AL.Get() parameters
+    /// A list of valid string AL.Get() parameters.
     /// </summary>
     public enum ALGetString
     {
