+++ src/OpenTK.OpenAL/Native/AL/Enums/ALSource3i.cs
@@ -1,7 +1,7 @@
 ﻿namespace OpenTK.OpenAL.Native
 {
     /// <summary>
-    /// A list of valid 3x Int32 Source/GetSource parameters
+    /// A list of valid 3x Int32 Source/GetSource parameters.
     /// </summary>
     public enum ALSource3i
     {
