+++ src/OpenTK.OpenAL/Native/AL/Enums/ALSourceType.cs
@@ -11,12 +11,12 @@
         Static = 0x1028,
 
         /// <summary>
-        /// Source is Streaming if one or more Buffers have been attached using AL.SourceQueueBuffers
+        /// Source is Streaming if one or more Buffers have been attached using AL.SourceQueueBuffers.
         /// </summary>
         Streaming = 0x1029,
 
         /// <summary>
-        /// Source is undetermined when it has a null Buffer attached
+        /// Source is undetermined when it has a null Buffer attached.
         /// </summary>
         Undetermined = 0x1030
     }
