+++ src/OpenTK.OpenAL/Native/AL/Extensions/Enums/EfxEffectf.cs
@@ -8,7 +8,7 @@
 namespace OpenTK.OpenAL.Native
 {
     /// <summary>
-    /// A list of valid 32-bit Float Effect/GetEffect parameters
+    /// A list of valid 32-bit Float Effect/GetEffect parameters.
     /// </summary>
     public enum EfxEffectf
     {
@@ -443,497 +443,5 @@ namespace OpenTK.OpenAL.Native
         /// Rolloff Factor, but operates on reverb sound instead of direct-path sound. Range [0.0f .. 10.0f] Default: 0.0f
         /// </summary>
         EaxReverbRoomRolloffFactor = 0x0016
-    }
-
-    /// <summary>
-    /// A list of valid Math.Vector3 Effect/GetEffect parameters
-    /// </summary>
-    public enum EfxEffect3f
-    {
-        /// <summary>
-        /// Reverb Pan does for the Reverb what Reflections Pan does for the Reflections. Unit: Vector3 of length 0f to 1f
-        /// Default: {0.0f, 0.0f, 0.0f}
-        /// </summary>
-        EaxReverbLateReverbPan = 0x000E,
-
-        /// <summary>
-        /// This Vector3 controls the spatial distribution of the cluster of early reflections. The direction of this
-        /// vector controls the global direction of the reflections, while its magnitude controls how focused the reflections
-        /// are towards this direction. For legacy reasons this Vector3 follows a left-handed co-ordinate system! Note that
-        /// OpenAL uses a right-handed coordinate system. Unit: Vector3 of length 0f to 1f Default: {0.0f, 0.0f, 0.0f}
-        /// </summary>
-        EaxReverbReflectionsPan = 0x000B
-    }
-
-    /// <summary>
-    /// A list of valid Int32 Effect/GetEffect parameters
-    /// </summary>
-    public enum EfxEffecti
-    {
-        /// <summary>
-        /// This property sets the waveform shape of the low-frequency oscillator that controls the delay time of the
-        /// delayed signals. Unit: (0) Sinusoid, (1) Triangle Range [0 .. 1] Default: 1
-        /// </summary>
-        ChorusWaveform = 0x0001,
-
-        /// <summary>
-        /// This property controls the phase difference between the left and right low-frequency oscillators. At zero
-        /// degrees the two low-frequency oscillators are synchronized. Unit: Degrees Range [-180 .. 180] Default: 90
-        /// </summary>
-        ChorusPhase = 0x0002,
-
-        /// <summary>
-        /// Selects the shape of the low-frequency oscillator waveform that controls the amount of the delay of the
-        /// sampled signal. Unit: (0) Sinusoid, (1) Triangle Range [0 .. 1] Default: 1
-        /// </summary>
-        FlangerWaveform = 0x0001,
-
-        /// <summary>
-        /// This changes the phase difference between the left and right low-frequency oscillator's. At zero degrees the
-        /// two low-frequency oscillators are synchronized. Range [-180 .. +180] Default: 0
-        /// </summary>
-        FlangerPhase = 0x0002,
-
-        /// <summary>
-        /// These select which internal signals are added together to produce the output. Unit: (0) Down, (1) Up, (2) Off
-        /// Range [0 .. 2] Default: 0
-        /// </summary>
-        FrequencyShifterLeftDirection = 0x0002,
-
-        /// <summary>
-        /// These select which internal signals are added together to produce the output. Unit: (0) Down, (1) Up, (2) Off
-        /// Range [0 .. 2] Default: 0
-        /// </summary>
-        FrequencyShifterRightDirection = 0x0003,
-
-        /// <summary>
-        /// Sets the vocal morpher 4-band formant filter A, used to impose vocal tract effects upon the input signal. The
-        /// vocal morpher is not necessarily intended for use on voice signals; it is primarily intended for pitched noise
-        /// effects, vocal-like wind effects, etc. Unit: Use enum EfxFormantFilterSettings Range [0 .. 29] Default: 0, "Phoneme
-        /// A"
-        /// </summary>
-        VocalMorpherPhonemeA = 0x0001,
-
-        /// <summary>
-        /// This is used to adjust the pitch of phoneme filter A in 1-semitone increments. Unit: Semitones Range [-24 ..
-        /// +24] Default: 0
-        /// </summary>
-        VocalMorpherPhonemeACoarseTuning = 0x0002,
-
-        /// <summary>
-        /// Sets the vocal morpher 4-band formant filter B, used to impose vocal tract effects upon the input signal. The
-        /// vocal morpher is not necessarily intended for use on voice signals; it is primarily intended for pitched noise
-        /// effects, vocal-like wind effects, etc. Unit: Use enum EfxFormantFilterSettings Range [0 .. 29] Default: 10,
-        /// "Phoneme ER"
-        /// </summary>
-        VocalMorpherPhonemeB = 0x0003,
-
-        /// <summary>
-        /// This is used to adjust the pitch of phoneme filter B in 1-semitone increments. Unit: Semitones Range [-24 ..
-        /// +24] Default: 0
-        /// </summary>
-        VocalMorpherPhonemeBCoarseTuning = 0x0004,
-
-        /// <summary>
-        /// This controls the shape of the low-frequency oscillator used to morph between the two phoneme filters. Unit:
-        /// (0) Sinusoid, (1) Triangle, (2) Sawtooth Range [0 .. 2] Default: 0
-        /// </summary>
-        VocalMorpherWaveform = 0x0005,
-
-        /// <summary>
-        /// This sets the number of semitones by which the pitch is shifted. There are 12 semitones per octave. Unit:
-        /// Semitones Range [-12 .. +12] Default: +12
-        /// </summary>
-        PitchShifterCoarseTune = 0x0001,
-
-        /// <summary>
-        /// This sets the number of cents between Semitones a pitch is shifted. A Cent is 1/100th of a Semitone. Unit:
-        /// Cents Range [-50 .. +50] Default: 0
-        /// </summary>
-        PitchShifterFineTune = 0x0002,
-
-        /// <summary>
-        /// This controls which waveform is used as the carrier signal. Traditional ring modulator and tremolo effects
-        /// generally use a sinusoidal carrier. Unit: (0) Sinusoid, (1) Sawtooth, (2) Square Range [0 .. 2] Default: 0
-        /// </summary>
-        RingModulatorWaveform = 0x0003,
-
-        /// <summary>
-        /// Enabling this will result in audio exhibiting smaller variation in intensity between the loudest and quietest
-        /// portions. Unit: (0) Off, (1) On Range [0 .. 1] Default: 1
-        /// </summary>
-        CompressorOnoff = 0x0001,
-
-        /// <summary>
-        /// When this flag is set, the high-frequency decay time automatically stays below a limit value that's derived
-        /// from the setting of the property Air Absorption HF. Unit: (0) False, (1) True Range [False, True] Default: True
-        /// </summary>
-        ReverbDecayHFLimit = 0x000D,
-
-        /// <summary>
-        /// When this flag is set, the high-frequency decay time automatically stays below a limit value that's derived
-        /// from the setting of the property AirAbsorptionGainHF. Unit: (0) False, (1) True Range [False, True] Default: True
-        /// </summary>
-        EaxReverbDecayHFLimit = 0x0017,
-
-        /// <summary>
-        /// Used with the enum EfxEffectType as it's parameter.
-        /// </summary>
-        EffectType = 0x8001
-    }
-
-    /// <summary>
-    /// Vocal morpher effect parameters. If both parameters are set to the same phoneme, that determines the filtering
-    /// effect that will be heard. If these two parameters are set to different phonemes, the filtering effect will morph
-    /// between the two settings at a rate specified by EfxEffectf.VocalMorpherRate.
-    /// </summary>
-    public enum EfxFormantFilterSettings
-    {
-        /// <summary>
-        /// The A phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeA = 0,
-
-        /// <summary>
-        /// The E phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeE = 1,
-
-        /// <summary>
-        /// The I phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeI = 2,
-
-        /// <summary>
-        /// The O phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeO = 3,
-
-        /// <summary>
-        /// The U phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeU = 4,
-
-        /// <summary>
-        /// The AA phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeAA = 5,
-
-        /// <summary>
-        /// The AE phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeAE = 6,
-
-        /// <summary>
-        /// The AH phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeAH = 7,
-
-        /// <summary>
-        /// The AO phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeAO = 8,
-
-        /// <summary>
-        /// The EH phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeEH = 9,
-
-        /// <summary>
-        /// The ER phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeER = 10,
-
-        /// <summary>
-        /// The IH phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeIH = 11,
-
-        /// <summary>
-        /// The IY phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeIY = 12,
-
-        /// <summary>
-        /// The UH phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeUH = 13,
-
-        /// <summary>
-        /// The UW phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeUW = 14,
-
-        /// <summary>
-        /// The B phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeB = 15,
-
-        /// <summary>
-        /// The D phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeD = 16,
-
-        /// <summary>
-        /// The F phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeF = 17,
-
-        /// <summary>
-        /// The G phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeG = 18,
-
-        /// <summary>
-        /// The J phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeJ = 19,
-
-        /// <summary>
-        /// The K phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeK = 20,
-
-        /// <summary>
-        /// The L phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeL = 21,
-
-        /// <summary>
-        /// The M phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeM = 22,
-
-        /// <summary>
-        /// The N phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeN = 23,
-
-        /// <summary>
-        /// The P phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeP = 24,
-
-        /// <summary>
-        /// The R phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeR = 25,
-
-        /// <summary>
-        /// The S phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeS = 26,
-
-        /// <summary>
-        /// The T phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeT = 27,
-
-        /// <summary>
-        /// The V phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeV = 28,
-
-        /// <summary>
-        /// The Z phoneme of the vocal morpher.
-        /// </summary>
-        VocalMorpherPhonemeZ = 29
-    }
-
-    /// <summary>
-    /// Effect type definitions to be used with EfxEffecti.EffectType.
-    /// </summary>
-    public enum EfxEffectType
-    {
-        /// <summary>
-        /// No Effect, disable. This Effect type is used when an Effect object is initially created.
-        /// </summary>
-        Null = 0x0000,
-
-        /// <summary>
-        /// The Reverb effect is the standard Effects Extension's environmental reverberation effect. It is available on
-        /// all Generic Software and Generic Hardware devices.
-        /// </summary>
-        Reverb = 0x0001,
-
-        /// <summary>
-        /// The Chorus effect essentially replays the input audio accompanied by another slightly delayed version of the
-        /// signal, creating a "doubling" effect.
-        /// </summary>
-        Chorus = 0x0002,
-
-        /// <summary>
-        /// The Distortion effect simulates turning up (overdriving) the gain stage on a guitar amplifier or adding a
-        /// distortion pedal to an instrument's output.
-        /// </summary>
-        Distortion = 0x0003,
-
-        /// <summary>
-        /// The Echo effect generates discrete, delayed instances of the input signal.
-        /// </summary>
-        Echo = 0x0004,
-
-        /// <summary>
-        /// The Flanger effect creates a "tearing" or "whooshing" sound, like a jet flying overhead.
-        /// </summary>
-        Flanger = 0x0005,
-
-        /// <summary>
-        /// The Frequency shifter is a single-sideband modulator, which translates all the component frequencies of the
-        /// input signal by an equal amount.
-        /// </summary>
-        FrequencyShifter = 0x0006,
-
-        /// <summary>
-        /// The Vocal morpher consists of a pair of 4-band formant filters, used to impose vocal tract effects upon the
-        /// input signal.
-        /// </summary>
-        VocalMorpher = 0x0007,
-
-        /// <summary>
-        /// The Pitch shifter applies time-invariant pitch shifting to the input signal, over a one octave range and
-        /// controllable at a semi-tone and cent resolution.
-        /// </summary>
-        PitchShifter = 0x0008,
-
-        /// <summary>
-        /// The Ring modulator multiplies an input signal by a carrier signal in the time domain, resulting in tremolo or
-        /// inharmonic effects.
-        /// </summary>
-        RingModulator = 0x0009,
-
-        /// <summary>
-        /// The Auto-wah effect emulates the sound of a wah-wah pedal used with an electric guitar, or a mute on a brass
-        /// instrument.
-        /// </summary>
-        Autowah = 0x000A,
-
-        /// <summary>
-        /// The Compressor will boost quieter portions of the audio, while louder portions will stay the same or may even
-        /// be reduced.
-        /// </summary>
-        Compressor = 0x000B,
-
-        /// <summary>
-        /// The Equalizer is very flexible, providing tonal control over four different adjustable frequency ranges.
-        /// </summary>
-        Equalizer = 0x000C,
-
-        /// <summary>
-        /// The EAX Reverb has a more advanced parameter set than EfxEffectType.Reverb, but is only natively supported on
-        /// devices that support the EAX 3.0 or above.
-        /// </summary>
-        EaxReverb = 0x8000
-    }
-
-    /// <summary>
-    /// A list of valid Int32 AuxiliaryEffectSlot/GetAuxiliaryEffectSlot parameters
-    /// </summary>
-    public enum EfxAuxiliaryi
-    {
-        /// <summary>
-        /// This property is used to attach an Effect object to the Auxiliary Effect Slot object. After the attachment,
-        /// the Auxiliary Effect Slot object will contain the effect type and have the same effect parameters that were stored
-        /// in the Effect object. Any Sources feeding the Auxiliary Effect Slot will immediate feed the new effect type and new
-        /// effect parameters.
-        /// </summary>
-        EffectslotEffect = 0x0001,
-
-        /// <summary>
-        /// This property is used to enable or disable automatic send adjustments based on the physical positions of the
-        /// sources and the listener. This property should be enabled when an application wishes to use a reverb effect to
-        /// simulate the environment surrounding a listener or a collection of Sources. Range [False, True] Default: True
-        /// </summary>
-        EffectslotAuxiliarySendAuto = 0x0003
-    }
-
-    /// <summary>
-    /// A list of valid 32-bits Float AuxiliaryEffectSlot/GetAuxiliaryEffectSlot parameters
-    /// </summary>
-    public enum EfxAuxiliaryf
-    {
-        /// <summary>
-        /// This property is used to specify an output level for the Auxiliary Effect Slot. Setting the gain to 0.0f mutes
-        /// the output. Range [0.0f .. 1.0f] Default: 1.0f
-        /// </summary>
-        EffectslotGain = 0x0002
-    }
-
-    /// <summary>
-    /// A list of valid 32-bits Float Filter/GetFilter parameters
-    /// </summary>
-    public enum EfxFilterf
-    {
-        /// <summary>
-        /// Range [0.0f .. 1.0f] Default: 1.0f
-        /// </summary>
-        LowpassGain = 0x0001,
-
-        /// <summary>
-        /// Range [0.0f .. 1.0f] Default: 1.0f
-        /// </summary>
-        LowpassGainHF = 0x0002,
-
-        /// <summary>
-        /// Range [0.0f .. 1.0f] Default: 1.0f
-        /// </summary>
-        HighpassGain = 0x0001,
-
-        /// <summary>
-        /// Range [0.0f .. 1.0f] Default: 1.0f
-        /// </summary>
-        HighpassGainLF = 0x0002,
-
-        /// <summary>
-        /// Range [0.0f .. 1.0f] Default: 1.0f
-        /// </summary>
-        BandpassGain = 0x0001,
-
-        /// <summary>
-        /// Range [0.0f .. 1.0f] Default: 1.0f
-        /// </summary>
-        BandpassGainLF = 0x0002,
-
-        /// <summary>
-        /// Range [0.0f .. 1.0f] Default: 1.0f
-        /// </summary>
-        BandpassGainHF = 0x0003
-    }
-
-    /// <summary>
-    /// A list of valid Int32 Filter/GetFilter parameters
-    /// </summary>
-    public enum EfxFilteri
-    {
-        /// <summary>
-        /// Used with the enum EfxFilterType as Parameter to select a filter logic.
-        /// </summary>
-        FilterType = 0x8001
-    }
-
-    /// <summary>
-    /// Filter type definitions to be used with EfxFilteri.FilterType.
-    /// </summary>
-    public enum EfxFilterType
-    {
-        /// <summary>
-        /// No Filter, disable. This Filter type is used when a Filter object is initially created.
-        /// </summary>
-        Null = 0x0000,
-
-        /// <summary>
-        /// A low-pass filter is used to remove high frequency content from a signal.
-        /// </summary>
-        Lowpass = 0x0001,
-
-        /// <summary>
-        /// Currently not implemented. A high-pass filter is used to remove low frequency content from a signal.
-        /// </summary>
-        Highpass = 0x0002,
-
-        /// <summary>
-        /// Currently not implemented. A band-pass filter is used to remove high and low frequency content from a signal.
-        /// </summary>
-        Bandpass = 0x0003
     }
 }
