+++ src/OpenTK.OpenAL/Native/AL/Extensions/Enums/EfxFilterf.cs
@@ -1,0 +1,43 @@
+﻿namespace OpenTK.OpenAL.Native
+{
+    /// <summary>
+    /// A list of valid 32-bits Float Filter/GetFilter parameters
+    /// </summary>
+    public enum EfxFilterf
+    {
+        /// <summary>
+        /// Range [0.0f .. 1.0f] Default: 1.0f
+        /// </summary>
+        LowpassGain = 0x0001,
+
+        /// <summary>
+        /// Range [0.0f .. 1.0f] Default: 1.0f
+        /// </summary>
+        LowpassGainHF = 0x0002,
+
+        /// <summary>
+        /// Range [0.0f .. 1.0f] Default: 1.0f
+        /// </summary>
+        HighpassGain = 0x0001,
+
+        /// <summary>
+        /// Range [0.0f .. 1.0f] Default: 1.0f
+        /// </summary>
+        HighpassGainLF = 0x0002,
+
+        /// <summary>
+        /// Range [0.0f .. 1.0f] Default: 1.0f
+        /// </summary>
+        BandpassGain = 0x0001,
+
+        /// <summary>
+        /// Range [0.0f .. 1.0f] Default: 1.0f
+        /// </summary>
+        BandpassGainLF = 0x0002,
+
+        /// <summary>
+        /// Range [0.0f .. 1.0f] Default: 1.0f
+        /// </summary>
+        BandpassGainHF = 0x0003
+    }
+}
