+++ src/OpenTK.OpenAL/Native/AL/Extensions/Enums/EfxFilteri.cs
@@ -1,0 +1,13 @@
+﻿namespace OpenTK.OpenAL.Native
+{
+    /// <summary>
+    /// A list of valid Int32 Filter/GetFilter parameters
+    /// </summary>
+    public enum EfxFilteri
+    {
+        /// <summary>
+        /// Used with the enum EfxFilterType as Parameter to select a filter logic.
+        /// </summary>
+        FilterType = 0x8001
+    }
+}
